Get location from one EcoCounter installation

Get location from one EcoCounter installation

get_EcoCounter_data(EcoCounterId, from = NULL, to = NULL, step = 4)

Arguments

EcoCounterId
character or integer (required): Id of the EcoCounter installation

from
character: Start date (YYYYMMDD). If no date is submitted, the day before yesterday is automatically set.

to
character: End date (YYYYMMDD). If no date is submitted, yesterday is automatically set. Note: When date is the date of today, usually no data are available.

step
integer or character: Get data : 2 (15 min), 3 (hourly), 4 (daily), 5 (weekly), 6 (monthly), 7 (yearly). With default: 4.

Value

data.frame with parsed data from EcoCounter API

Function version 0.0.1

NA

Examples

EcoCounterId <- 101020207
get_EcoCounter_data(EcoCounterId)

#>         date comptage        id
#> 1 2018-01-30     9802 101020207
#> 2 2018-01-31     6111 101020207

## set specific date and request weekly data (step = 5)
get_EcoCounter_data(EcoCounterId, from = "20171101", to = "20171201", step = 5)

#>         date comptage        id
#> 1 2017-11-01    33404 101020207
#> 2 2017-11-06    63290 101020207
#> 3 2017-11-13    65443 101020207
#> 4 2017-11-20    77073 101020207
#> 5 2017-11-27    69403 101020207