GET info about SpaceX capsules

GET info about SpaceX capsules

get_SpaceX_capsules(type = NULL, ...)

Arguments

type
character: type of the request: Choose between "upcoming" or "past".

...
: Query string to API. See more in details.

Value

data.frame with parsed data from SpaceX API.

Examples

# get all capsules
get_SpaceX_capsules()

#>    capsule_serial capsule_id    status          original_launch
#> 1            C101    dragon1   retired 2010-12-08T15:43:00.000Z
#> 2            C102    dragon1   retired 2012-05-22T07:44:00.000Z
#> 3            C103    dragon1   unknown 2012-10-08T00:35:00.000Z
#> 4            C104    dragon1   unknown 2013-03-01T19:10:00.000Z
#> 5            C105    dragon1   unknown 2014-04-18T19:25:00.000Z
#> 6            C106    dragon1    active 2014-09-21T05:52:00.000Z
#> 7            C107    dragon1   unknown 2015-01-10T09:47:00.000Z
#> 8            C108    dragon1    active 2015-04-14T20:10:00.000Z
#> 9            C109    dragon1 destroyed 2015-06-28T14:21:00.000Z
#> 10           C110    dragon1    active 2016-04-08T20:43:00.000Z
#> 11           C111    dragon1    active 2016-07-18T04:45:00.000Z
#> 12           C112    dragon1    active 2017-02-19T14:39:00.000Z
#> 13           C113    dragon1    active 2017-08-14T16:31:00.000Z
#> 14           C201    dragon2    active 2019-03-02T07:45:00.000Z
#> 15           C205    dragon2    active 2020-01-19T14:00:00.000Z
#> 16           C206    dragon2    active 2020-05-30T19:22:00.000Z
#> 17           C202    dragon2    active                     <NA>
#> 18           C203    dragon2    active                     <NA>
#> 19           C204    dragon2    active                     <NA>
#>    original_launch_unix                             missions landings
#> 1            1291822980                            COTS 1, 7        0
#> 2            1335944640                            COTS 2, 8        1
#> 3            1349656500                             CRS-1, 9        1
#> 4            1362165000                            CRS-2, 10        1
#> 5            1397849100                            CRS-3, 14        1
#> 6            1411278720    CRS-4, CRS-11, CRS-19, 18, 41, 85        3
#> 7            1420883220                            CRS-5, 19        1
#> 8            1429042200    CRS-6, CRS-13, CRS-18, 22, 51, 82        3
#> 9            1435501260                            CRS-7, 24        0
#> 10           1460148180                CRS-8, CRS-14, 28, 59        2
#> 11           1468817100                CRS-9, CRS-15, 32, 64        2
#> 12           1487515140   CRS-10, CRS-16, CRS-20, 36, 72, 91        3
#> 13           1502728260               CRS-12, CRS-17, 45, 78        2
#> 14           1551512700            CCtCap Demo Mission 1, 76        1
#> 15           1579442400 Crew Dragon In Flight Abort Test, 88        1
#> 16           1590866520            CCtCap Demo Mission 2, 94        1
#> 17                   NA                                 NULL        0
#> 18                   NA                                 NULL        0
#> 19                   NA                                 NULL        0
#>          type
#> 1  Dragon 1.0
#> 2  Dragon 1.0
#> 3  Dragon 1.0
#> 4  Dragon 1.0
#> 5  Dragon 1.1
#> 6  Dragon 1.1
#> 7  Dragon 1.1
#> 8  Dragon 1.1
#> 9  Dragon 1.1
#> 10 Dragon 1.1
#> 11 Dragon 1.1
#> 12 Dragon 1.1
#> 13 Dragon 1.1
#> 14 Dragon 2.0
#> 15 Dragon 2.0
#> 16 Dragon 2.0
#> 17 Dragon 2.0
#> 18 Dragon 2.0
#> 19 Dragon 2.0
#>                                                                                                  details
#> 1                                                                   Reentered after three weeks in orbit
#> 2                                                                                First Dragon spacecraft
#> 3                                                 First of twenty missions flown under the CRS1 contract
#> 4                                                                                                   <NA>
#> 5                                                                              First Dragon v1.1 capsule
#> 6                                                                      First Dragon capsule to be reused
#> 7                                                                                                   <NA>
#> 8                                                                     Second Dragon capsule to be reused
#> 9                                                            Destroyed on impact after F9 launch failure
#> 10                                                                                                  <NA>
#> 11                                                                                                  <NA>
#> 12                                                                                                  <NA>
#> 13                                                                  The last newly manufactured Dragon 1
#> 14                                                 Destroyed in a test anomaly at LZ-1 on April 20, 2019
#> 15                        In construction for use in first mission in contract under the CCtCap contract
#> 16 In Hawthorne, CA as of October 11, 2019. Seen in background during Musk/Bridenstine press conference.
#> 17                    Capsule used to qualify Dragon 2's environmental control and life support systems.
#> 18                                                 Rumored to be used for Inflight Abort Test after DM-1
#> 19                                                             Currently in construction for use in DM-2
#>    reuse_count
#> 1            0
#> 2            0
#> 3            0
#> 4            0
#> 5            0
#> 6            2
#> 7            0
#> 8            2
#> 9            0
#> 10           1
#> 11           1
#> 12           2
#> 13           1
#> 14           0
#> 15           0
#> 16           0
#> 17           0
#> 18           0
#> 19           0

# get past capsules
get_SpaceX_capsules("past")

#>    capsule_serial capsule_id    status          original_launch
#> 1            C101    dragon1   retired 2010-12-08T15:43:00.000Z
#> 2            C102    dragon1   retired 2012-05-22T07:44:00.000Z
#> 3            C103    dragon1   unknown 2012-10-08T00:35:00.000Z
#> 4            C104    dragon1   unknown 2013-03-01T19:10:00.000Z
#> 5            C105    dragon1   unknown 2014-04-18T19:25:00.000Z
#> 6            C106    dragon1    active 2014-09-21T05:52:00.000Z
#> 7            C107    dragon1   unknown 2015-01-10T09:47:00.000Z
#> 8            C108    dragon1    active 2015-04-14T20:10:00.000Z
#> 9            C109    dragon1 destroyed 2015-06-28T14:21:00.000Z
#> 10           C110    dragon1    active 2016-04-08T20:43:00.000Z
#> 11           C111    dragon1    active 2016-07-18T04:45:00.000Z
#> 12           C112    dragon1    active 2017-02-19T14:39:00.000Z
#> 13           C113    dragon1    active 2017-08-14T16:31:00.000Z
#> 14           C201    dragon2    active 2019-03-02T07:45:00.000Z
#> 15           C205    dragon2    active 2020-01-19T14:00:00.000Z
#> 16           C206    dragon2    active 2020-05-30T19:22:00.000Z
#>    original_launch_unix                             missions landings
#> 1            1291822980                            COTS 1, 7        0
#> 2            1335944640                            COTS 2, 8        1
#> 3            1349656500                             CRS-1, 9        1
#> 4            1362165000                            CRS-2, 10        1
#> 5            1397849100                            CRS-3, 14        1
#> 6            1411278720    CRS-4, CRS-11, CRS-19, 18, 41, 85        3
#> 7            1420883220                            CRS-5, 19        1
#> 8            1429042200    CRS-6, CRS-13, CRS-18, 22, 51, 82        3
#> 9            1435501260                            CRS-7, 24        0
#> 10           1460148180                CRS-8, CRS-14, 28, 59        2
#> 11           1468817100                CRS-9, CRS-15, 32, 64        2
#> 12           1487515140   CRS-10, CRS-16, CRS-20, 36, 72, 91        3
#> 13           1502728260               CRS-12, CRS-17, 45, 78        2
#> 14           1551512700            CCtCap Demo Mission 1, 76        1
#> 15           1579442400 Crew Dragon In Flight Abort Test, 88        1
#> 16           1590866520            CCtCap Demo Mission 2, 94        1
#>          type
#> 1  Dragon 1.0
#> 2  Dragon 1.0
#> 3  Dragon 1.0
#> 4  Dragon 1.0
#> 5  Dragon 1.1
#> 6  Dragon 1.1
#> 7  Dragon 1.1
#> 8  Dragon 1.1
#> 9  Dragon 1.1
#> 10 Dragon 1.1
#> 11 Dragon 1.1
#> 12 Dragon 1.1
#> 13 Dragon 1.1
#> 14 Dragon 2.0
#> 15 Dragon 2.0
#> 16 Dragon 2.0
#>                                                                                                  details
#> 1                                                                   Reentered after three weeks in orbit
#> 2                                                                                First Dragon spacecraft
#> 3                                                 First of twenty missions flown under the CRS1 contract
#> 4                                                                                                   <NA>
#> 5                                                                              First Dragon v1.1 capsule
#> 6                                                                      First Dragon capsule to be reused
#> 7                                                                                                   <NA>
#> 8                                                                     Second Dragon capsule to be reused
#> 9                                                            Destroyed on impact after F9 launch failure
#> 10                                                                                                  <NA>
#> 11                                                                                                  <NA>
#> 12                                                                                                  <NA>
#> 13                                                                  The last newly manufactured Dragon 1
#> 14                                                 Destroyed in a test anomaly at LZ-1 on April 20, 2019
#> 15                        In construction for use in first mission in contract under the CCtCap contract
#> 16 In Hawthorne, CA as of October 11, 2019. Seen in background during Musk/Bridenstine press conference.
#>    reuse_count
#> 1            0
#> 2            0
#> 3            0
#> 4            0
#> 5            0
#> 6            2
#> 7            0
#> 8            2
#> 9            0
#> 10           1
#> 11           1
#> 12           2
#> 13           1
#> 14           0
#> 15           0
#> 16           0