GET info about SpaceX all dragon capsules

GET info about SpaceX all dragon capsules

get_SpaceX_dragons(...)

Arguments

...
: Query string to API. See more in details.

Value

data.frame with parsed data from SpaceX API.

Examples

# get all data for dragon capsules
get_SpaceX_dragons()

#>        id     name    type active crew_capacity sidewall_angle_deg
#> 1 dragon1 Dragon 1 capsule   TRUE             0                 15
#> 2 dragon2 Dragon 2 capsule   TRUE             7                 15
#>   orbit_duration_yr dry_mass_kg dry_mass_lb first_flight heat_shield.material
#> 1                 2        4200        9300   2010-12-08               PICA-X
#> 2                 2        6350       14000   2019-03-02               PICA-X
#>   heat_shield.size_meters heat_shield.temp_degrees heat_shield.dev_partner
#> 1                     3.6                     3000                    NASA
#> 2                     3.6                     3000                    NASA
#>                                                                                                                                          thrusters
#> 1                                                                              Draco, 18, 4, nitrogen tetroxide, monomethylhydrazine, 300, 0.4, 90
#> 2 Draco, SuperDraco, 18, 8, 4, 4, nitrogen tetroxide, dinitrogen tetroxide, monomethylhydrazine, monomethylhydrazine, 300, 235, 0.4, 71, 90, 16000
#>   launch_payload_mass.kg launch_payload_mass.lb launch_payload_vol.cubic_meters
#> 1                   6000                  13228                              25
#> 2                   6000                  13228                              25
#>   launch_payload_vol.cubic_feet return_payload_mass.kg return_payload_mass.lb
#> 1                           883                   3000                   6614
#> 2                           883                   3000                   6614
#>   return_payload_vol.cubic_meters return_payload_vol.cubic_feet
#> 1                              11                           388
#> 2                              11                           388
#>   pressurized_capsule.payload_volume.cubic_meters
#> 1                                              11
#> 2                                              11
#>   pressurized_capsule.payload_volume.cubic_feet trunk.trunk_volume.cubic_meters
#> 1                                           388                              14
#> 2                                           388                              14
#>   trunk.trunk_volume.cubic_feet trunk.cargo.solar_array
#> 1                           494                       2
#> 2                           494                       2
#>   trunk.cargo.unpressurized_cargo height_w_trunk.meters height_w_trunk.feet
#> 1                            TRUE                   7.2                23.6
#> 2                            TRUE                   7.2                23.6
#>   diameter.meters diameter.feet
#> 1             3.7            12
#> 2             3.7            12
#>                                                                                                                                                                                                                                                                                                                                                                                    flickr_images
#> 1 https://www.spacex.com/sites/spacex/files/styles/media_gallery_large/public/2015_-_04_crs5_dragon_orbit13.jpg?itok=9p8_l7UP, https://www.spacex.com/sites/spacex/files/styles/media_gallery_large/public/2012_-_4_dragon_grapple_cots2-1.jpg?itok=R2-SeuMX, https://farm3.staticflickr.com/2815/32761844973_4b55b27d3c_b.jpg, https://farm9.staticflickr.com/8618/16649075267_d18cbb4342_b.jpg
#> 2                                                                                                                                                                                            https://farm8.staticflickr.com/7647/16581815487_6d56cb32e1_b.jpg, https://farm1.staticflickr.com/780/21119686299_c88f63e350_b.jpg, https://farm9.staticflickr.com/8588/16661791299_a236e2f5dc_b.jpg
#>                                     wikipedia
#> 1 https://en.wikipedia.org/wiki/SpaceX_Dragon
#> 2      https://en.wikipedia.org/wiki/Dragon_2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                description
#> 1                                                                                                                                                   Dragon is a reusable spacecraft developed by SpaceX, an American private space transportation company based in Hawthorne, California. Dragon is launched into space by the SpaceX Falcon 9 two-stage-to-orbit launch vehicle. The Dragon spacecraft was originally designed for human travel, but so far has only been used to deliver cargo to the International Space Station (ISS).
#> 2 Dragon 2 (also Crew Dragon, Dragon V2, or formerly DragonRider) is the second version of the SpaceX Dragon spacecraft, which will be a human-rated vehicle. It includes a set of four side-mounted thruster pods with two SuperDraco engines each, which can serve as a launch escape system or launch abort system (LAS). In addition, it has much larger windows, new flight computers and avionics, and redesigned solar arrays, and a modified outer mold line from the initial cargo Dragon that has been flying for several years.

# get data for "dragon1"
get_SpaceX_dragons(id = "dragon1")

#>        id     name    type active crew_capacity sidewall_angle_deg
#> 1 dragon1 Dragon 1 capsule   TRUE             0                 15
#> 2 dragon2 Dragon 2 capsule   TRUE             7                 15
#>   orbit_duration_yr dry_mass_kg dry_mass_lb first_flight heat_shield.material
#> 1                 2        4200        9300   2010-12-08               PICA-X
#> 2                 2        6350       14000   2019-03-02               PICA-X
#>   heat_shield.size_meters heat_shield.temp_degrees heat_shield.dev_partner
#> 1                     3.6                     3000                    NASA
#> 2                     3.6                     3000                    NASA
#>                                                                                                                                          thrusters
#> 1                                                                              Draco, 18, 4, nitrogen tetroxide, monomethylhydrazine, 300, 0.4, 90
#> 2 Draco, SuperDraco, 18, 8, 4, 4, nitrogen tetroxide, dinitrogen tetroxide, monomethylhydrazine, monomethylhydrazine, 300, 235, 0.4, 71, 90, 16000
#>   launch_payload_mass.kg launch_payload_mass.lb launch_payload_vol.cubic_meters
#> 1                   6000                  13228                              25
#> 2                   6000                  13228                              25
#>   launch_payload_vol.cubic_feet return_payload_mass.kg return_payload_mass.lb
#> 1                           883                   3000                   6614
#> 2                           883                   3000                   6614
#>   return_payload_vol.cubic_meters return_payload_vol.cubic_feet
#> 1                              11                           388
#> 2                              11                           388
#>   pressurized_capsule.payload_volume.cubic_meters
#> 1                                              11
#> 2                                              11
#>   pressurized_capsule.payload_volume.cubic_feet trunk.trunk_volume.cubic_meters
#> 1                                           388                              14
#> 2                                           388                              14
#>   trunk.trunk_volume.cubic_feet trunk.cargo.solar_array
#> 1                           494                       2
#> 2                           494                       2
#>   trunk.cargo.unpressurized_cargo height_w_trunk.meters height_w_trunk.feet
#> 1                            TRUE                   7.2                23.6
#> 2                            TRUE                   7.2                23.6
#>   diameter.meters diameter.feet
#> 1             3.7            12
#> 2             3.7            12
#>                                                                                                                                                                                                                                                                                                                                                                                    flickr_images
#> 1 https://www.spacex.com/sites/spacex/files/styles/media_gallery_large/public/2015_-_04_crs5_dragon_orbit13.jpg?itok=9p8_l7UP, https://www.spacex.com/sites/spacex/files/styles/media_gallery_large/public/2012_-_4_dragon_grapple_cots2-1.jpg?itok=R2-SeuMX, https://farm3.staticflickr.com/2815/32761844973_4b55b27d3c_b.jpg, https://farm9.staticflickr.com/8618/16649075267_d18cbb4342_b.jpg
#> 2                                                                                                                                                                                            https://farm8.staticflickr.com/7647/16581815487_6d56cb32e1_b.jpg, https://farm1.staticflickr.com/780/21119686299_c88f63e350_b.jpg, https://farm9.staticflickr.com/8588/16661791299_a236e2f5dc_b.jpg
#>                                     wikipedia
#> 1 https://en.wikipedia.org/wiki/SpaceX_Dragon
#> 2      https://en.wikipedia.org/wiki/Dragon_2
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                description
#> 1                                                                                                                                                   Dragon is a reusable spacecraft developed by SpaceX, an American private space transportation company based in Hawthorne, California. Dragon is launched into space by the SpaceX Falcon 9 two-stage-to-orbit launch vehicle. The Dragon spacecraft was originally designed for human travel, but so far has only been used to deliver cargo to the International Space Station (ISS).
#> 2 Dragon 2 (also Crew Dragon, Dragon V2, or formerly DragonRider) is the second version of the SpaceX Dragon spacecraft, which will be a human-rated vehicle. It includes a set of four side-mounted thruster pods with two SuperDraco engines each, which can serve as a launch escape system or launch abort system (LAS). In addition, it has much larger windows, new flight computers and avionics, and redesigned solar arrays, and a modified outer mold line from the initial cargo Dragon that has been flying for several years.