GET info about SpaceX history

GET info about SpaceX history

get_SpaceX_history(...)

Arguments

...
: Query string to API. See more in details.

Value

data.frame with parsed data from SpaceX API.

Examples

# get all data
get_SpaceX_history()

#>    id                                      title           event_date_utc
#> 1   1                     Falcon 1 Makes History     2008-09-28T23:15:00Z
#> 2   2             SpaceX Wins $1.6B CRS Contract     2008-12-23T01:00:00Z
#> 3   3            Falcon 1 Flight 5 Makes History     2009-07-13T03:35:00Z
#> 4   4                      Falcon 9 First Flight     2010-06-04T18:45:00Z
#> 5   5            Dragon Returns From Earth Orbit     2010-12-08T15:43:00Z
#> 6   6        First Dragon Visit to Space Station     2012-10-08T00:35:00Z
#> 7   7     Grasshopper Completes Half-Mile Flight     2013-10-07T01:00:00Z
#> 8   8            First Flight of Falcon 9 to GTO     2013-12-03T22:41:00Z
#> 9   9                        First Stage Landing     2014-04-18T19:25:00Z
#> 10 10 Falcon 9 Reusable Test Vehicle Flies 1000M     2014-05-02T01:00:00Z
#> 11 11    SpaceX Awarded Commercial Crew Contract     2014-09-16T01:00:00Z
#> 12 12                        Drone Ship Landings     2015-01-10T09:47:00Z
#> 13 13                             Pad Abort Test     2015-05-06T13:00:00Z
#> 14 14                         First Land Landing     2015-12-22T01:29:00Z
#> 15 15                          Droneship Landing     2016-04-08T20:43:00Z
#> 16 16                             First Reflight     2017-03-30T22:27:00Z
#> 17 17                      First Dragon Reflight     2017-06-03T21:07:00Z
#> 18 18                   Falcon Heavy Test Flight     2018-02-06T20:45:00Z
#> 19 19                       First Block 5 Launch     2018-05-11T20:14:00Z
#> 20 20                     First Block 5 Reflight 2018-08-07T05:18:00.000Z
#>    event_date_unix flight_number
#> 1       1222643700             4
#> 2       1229994000            NA
#> 3       1247456100             5
#> 4       1275677100             6
#> 5       1291822980             7
#> 6       1349656500             9
#> 7       1381107600            NA
#> 8       1386110460            12
#> 9       1397849100            14
#> 10      1398992400            NA
#> 11      1410829200            NA
#> 12      1420883220            19
#> 13      1430917200            NA
#> 14      1450747740            25
#> 15      1460148180            28
#> 16      1490912820            38
#> 17      1496524020            41
#> 18      1517949900            55
#> 19      1526069640            61
#> 20      1533619080            67
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                details
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              Falcon 1 becomes the first privately developed liquid fuel rocket to reach Earth orbit.
#> 2                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                NASA awards SpaceX $1.6B Commercial Resupply Services (CRS) contract.
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         Falcon 1 Flight 5 makes history, becoming the first privately developed liquid fuel rocket to deliver a commercial satellite to Earth orbit.
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                  Met 100% of mission objectives on the first flight!
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             On December 8, 2010, Dragon became the first privately developed spacecraft in history to re-enter from low-Earth orbit.
#> 6                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                   Dragon becomes the first private spacecraft in history to visit the space station.
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                             The Grasshopper program finished with a 744m flight, hover, and landing.
#> 8                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       Falcon 9 reaches Geosynchronous Transfer Orbit
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                           Falcon 9 first stage successfully lands in Atlantic Ocean.
#> 10                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               Vehicle completes highest leap to date, lands safely.
#> 11                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                NASA awards $2.6 billion SpaceX contract to fly American astronauts.
#> 12                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         SpaceX begins series of first stage landing attempts on an autonomous spaceport drone ship.
#> 13                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    Crew Dragon tests launch abort system, which can provide astronauts with escape capability all the way to orbit.
#> 14                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                On December 21, 2015, the Falcon 9 rocket delivered 11 communications satellites to orbit, and the first stage returned and landed at Landing Zone 1 -– the first-ever orbital class rocket landing.
#> 15                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                       On April 8, 2016, the Falcon 9 rocket launched the Dragon spacecraft to the International Space Station, and the first stage returned and landed on the Of Course I Still Love You droneship.
#> 16                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                        On March 30, 2017, SpaceX achieved the world's first reflight of an orbital class rocket. Following delivery of the payload, the Falcon 9 first stage returned to Earth for the second time.
#> 17                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 This Dragon resupply mission represented the first reflight of a commercial spacecraft to and from the International Space Station.
#> 18 Maiden flight of Falcon Heavy, using two recovered Falcon 9 cores as side boosters (from the Thaicom 8 and CRS-9 missions), as well as a modified Block 3 booster reinforced to endure the additional load from the two side boosters. The static fire test, held on January 24, was the first time 27 engines were tested together. The launch was a success, and the side boosters landed simultaneously at adjacent ground pads. Drone ship landing of the central core failed due to TEA–TEB chemical igniter running out, preventing two of its engines from restarting; the landing failure caused damage to the nearby drone ship. Final burn to heliocentric Mars–Earth orbit was performed after the second stage and payload cruised for 6 hours through the Van Allen belts. Later, Elon Musk tweeted that the third burn was successful, and JPL's HORIZONS system showed the second stage and payload in an orbit with an aphelion of 1.67 AU. The live webcast proved immensely popular, as it became the second most watched livestream ever on YouTube, reaching over 2.3 million concurrent views.
#> 19                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         The Bangabandhu Satellite-1 mission served as the first flight of Falcon 9 Block 5, the final substantial upgrade to SpaceX’s Falcon 9 launch vehicle. Falcon 9 Block 5 is designed to be capable of 10 or more flights with very limited refurbishment as SpaceX continues to strive for rapid reusability and extremely high reliability.
#> 20                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                    Indonesian comsat intended to replace the aging Telkom 1 at 108° E. First reflight of a Block 5-version booster.
#>                                                                                         links.reddit
#> 1                                                                                               <NA>
#> 2                                                                                               <NA>
#> 3                                                                                               <NA>
#> 4                                                                                               <NA>
#> 5                                                                                               <NA>
#> 6                                                                                               <NA>
#> 7                                                                                               <NA>
#> 8                                                                                               <NA>
#> 9                                                                                               <NA>
#> 10                                                                                              <NA>
#> 11                                                                                              <NA>
#> 12                                                    http://www.reddit.com/r/spacex/comments/2rrdha
#> 13               https://www.reddit.com/r/spacex/comments/3527zv/official_video_pad_abort_test_2015/
#> 14     https://www.reddit.com/r/spacex/comments/3xgxh5/rspacex_orbcomm_og2_launch_2_official_launch/
#> 15   https://www.reddit.com/r/spacex/comments/4dtoly/rspacex_spacex_crs8_official_launch_discussion/
#> 16 https://www.reddit.com/r/spacex/comments/62aqi7/rspacex_ses10_official_launch_discussion_updates/
#> 17     https://www.reddit.com/r/spacex/comments/6ektkt/welcome_to_the_rspacex_crs11_official_launch/
#> 18 https://www.reddit.com/r/spacex/comments/7vg63x/rspacex_falcon_heavy_test_flight_official_launch/
#> 19  https://www.reddit.com/r/spacex/comments/8ioq3u/rspacex_bangabandhu1_official_launch_discussion/
#> 20      https://www.reddit.com/r/spacex/comments/9539nr/rspacex_merah_putih_telkom4_official_launch/
#>                                                                                                                           links.article
#> 1                                                                        http://www.spacex.com/news/2013/02/11/flight-4-launch-update-0
#> 2                                           https://www.nasaspaceflight.com/2008/12/spacex-and-orbital-win-huge-crs-contract-from-nasa/
#> 3                                                                               http://www.spacex.com/news/2013/02/12/falcon-1-flight-5
#> 4                                                                                                      http://www.bbc.com/news/10209704
#> 5                                                                              http://www.cnn.com/2010/US/12/08/space.flight/index.html
#> 6                                                                                              http://thespacereview.com/article/2168/1
#> 7                                                 http://spacenews.com/37740spacex-retires-grasshopper-new-test-rig-to-fly-in-december/
#> 8                                               http://www.newspacejournal.com/2013/03/27/after-dragon-spacexs-focus-returns-to-falcon/
#> 9                                                                 https://www.space.com/25562-spacex-falcon-9-reusable-rocket-test.html
#> 10                                                  http://www.parabolicarc.com/2014/05/02/falcon-9-reusable-vehicle-flies-1000-meters/
#> 11 https://www.washingtonpost.com/news/the-switch/wp/2014/09/16/nasa-awards-space-contract-to-boeing-and-spacex/?utm_term=.d6388390d071
#> 12                                 https://spaceflightnow.com/2015/01/10/dragon-successfully-launched-rocket-recovery-demo-crash-lands/
#> 13                                                       https://spaceflightnow.com/2015/04/21/dragon-pad-abort-test-set-for-early-may/
#> 14                                 https://spaceflightnow.com/2015/12/22/round-trip-rocket-flight-gives-spacex-a-trifecta-of-successes/
#> 15                        https://spaceflightnow.com/2016/04/08/spacex-lands-rocket-on-floating-platform-after-station-resupply-launch/
#> 16               https://spaceflightnow.com/2017/03/31/spacex-flies-rocket-for-second-time-in-historic-test-of-cost-cutting-technology/
#> 17                              https://spaceflightnow.com/2017/06/03/reused-dragon-cargo-capsule-launched-on-journey-to-space-station/
#> 18                https://spaceflightnow.com/2018/02/07/spacex-debuts-worlds-most-powerful-rocket-sends-tesla-toward-the-asteroid-belt/
#> 19                            https://spaceflightnow.com/2018/05/11/spacex-debuts-an-improved-human-rated-model-of-the-falcon-9-rocket/
#> 20                               https://spaceflightnow.com/2018/08/07/indonesian-communications-satellite-deployed-in-orbit-by-spacex/
#>                                                       links.wikipedia
#> 1                              https://en.wikipedia.org/wiki/Falcon_1
#> 2          https://en.wikipedia.org/wiki/Commercial_Resupply_Services
#> 3                              https://en.wikipedia.org/wiki/Falcon_1
#> 4  https://en.wikipedia.org/wiki/Dragon_Spacecraft_Qualification_Unit
#> 5             https://en.wikipedia.org/wiki/SpaceX_COTS_Demo_Flight_1
#> 6                          https://en.wikipedia.org/wiki/SpaceX_CRS-1
#> 7                  https://en.wikipedia.org/wiki/Grasshopper_(rocket)
#> 8                                 https://en.wikipedia.org/wiki/SES-8
#> 9                          https://en.wikipedia.org/wiki/SpaceX_CRS-3
#> 10                                                               <NA>
#> 11                                                               <NA>
#> 12                         https://en.wikipedia.org/wiki/SpaceX_CRS-5
#> 13                       https://en.wikipedia.org/wiki/Pad_abort_test
#> 14                   https://en.wikipedia.org/wiki/Falcon_9_flight_20
#> 15                         https://en.wikipedia.org/wiki/SpaceX_CRS-8
#> 16                               https://en.wikipedia.org/wiki/SES-10
#> 17                        https://en.wikipedia.org/wiki/SpaceX_CRS-11
#> 18             https://en.wikipedia.org/wiki/Falcon_Heavy_test_flight
#> 19                        https://en.wikipedia.org/wiki/Bangabandhu-1
#> 20                     https://en.wikipedia.org/wiki/Telkom_Indonesia

get_SpaceX_history(id = 17)

#>   id                 title       event_date_utc event_date_unix flight_number
#> 1 17 First Dragon Reflight 2017-06-03T21:07:00Z      1496524020            41
#>                                                                                                                               details
#> 1 This Dragon resupply mission represented the first reflight of a commercial spacecraft to and from the International Space Station.
#>                                                                                    links.reddit
#> 1 https://www.reddit.com/r/spacex/comments/6ektkt/welcome_to_the_rspacex_crs11_official_launch/
#>                                                                                             links.article
#> 1 https://spaceflightnow.com/2017/06/03/reused-dragon-cargo-capsule-launched-on-journey-to-space-station/
#>                               links.wikipedia
#> 1 https://en.wikipedia.org/wiki/SpaceX_CRS-11

# get data with flight_number 41
get_SpaceX_history(flight_number = 41)

#>   id                 title       event_date_utc event_date_unix flight_number
#> 1 17 First Dragon Reflight 2017-06-03T21:07:00Z      1496524020            41
#>                                                                                                                               details
#> 1 This Dragon resupply mission represented the first reflight of a commercial spacecraft to and from the International Space Station.
#>                                                                                    links.reddit
#> 1 https://www.reddit.com/r/spacex/comments/6ektkt/welcome_to_the_rspacex_crs11_official_launch/
#>                                                                                             links.article
#> 1 https://spaceflightnow.com/2017/06/03/reused-dragon-cargo-capsule-launched-on-journey-to-space-station/
#>                               links.wikipedia
#> 1 https://en.wikipedia.org/wiki/SpaceX_CRS-11