GET info about SpaceX company

GET info about SpaceX company

get_SpaceX_info()

Examples

get_SpaceX_info()

#>     name   founder founded employees vehicles launch_sites test_sites       ceo
#> 1 SpaceX Elon Musk    2002      7000        3            3          1 Elon Musk
#>         cto             coo cto_propulsion valuation headquarters.address
#> 1 Elon Musk Gwynne Shotwell    Tom Mueller   1.5e+10          Rocket Road
#>   headquarters.city headquarters.state
#> 1         Hawthorne         California
#>                                                                                                                                                                                                             summary
#> 1 SpaceX designs, manufactures and launches advanced rockets and spacecraft. The company was founded in 2002 to revolutionize space technology, with the ultimate goal of enabling people to live on other planets.