GET info about SpaceX launchpads

GET info about SpaceX launchpads

get_SpaceX_launchpads(name = NULL)

Arguments

name
character: Name of lauchnpads. Choose "ksc_lc_39a" or NULL. By default all three rocket data are returned.

Value

data.frame with parsed data from SpaceX API.

Examples

## get details about all launchpads
get_SpaceX_launchpads()

#>                id                                                full_name
#> 1    ccafs_slc_40 Cape Canaveral Air Force Station Space Launch Complex 40
#> 2            stls                           SpaceX South Texas Launch Site
#> 3     vafb_slc_4w        Vandenberg Air Force Base Space Launch Complex 4W
#> 4     ccafs_lc_13 Cape Canaveral Air Force Station Space Launch Complex 13
#> 5 kwajalein_atoll                            Kwajalein Atoll Omelek Island
#> 6      ksc_lc_39a         Kennedy Space Center Historic Launch Complex 39A
#> 7     vafb_slc_3w        Vandenberg Air Force Base Space Launch Complex 3W
#> 8     vafb_slc_4e        Vandenberg Air Force Base Space Launch Complex 4E
#>               status             location.name  location.region
#> 1             active            Cape Canaveral          Florida
#> 2 under construction        Boca Chica Village            Texas
#> 3             active Vandenberg Air Force Base       California
#> 4             active            Cape Canaveral          Florida
#> 5            retired             Omelek Island Marshall Islands
#> 6             active            Cape Canaveral          Florida
#> 7            retired Vandenberg Air Force Base       California
#> 8             active Vandenberg Air Force Base       California
#>   location.latitude location.longitude      vehicles_launched
#> 1         28.561857          -80.57737               Falcon 9
#> 2         25.997264          -97.15608               Falcon 9
#> 3         34.633204         -120.61562               Falcon 9
#> 4         28.485724          -80.54492               Falcon 9
#> 5          9.047721          167.74313               Falcon 1
#> 6         28.608058          -80.60396 Falcon 9, Falcon Heavy
#> 7         34.644090         -120.59314               Falcon 9
#> 8         34.632093         -120.61083               Falcon 9
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                    details
#> 1                                                                                                                                                                                                                                             SpaceX primary Falcon 9 launch pad, where all east coast Falcon 9s launched prior to the AMOS-6 anomaly. Initially used to launch Titan rockets for Lockheed Martin. Back online since CRS-13 on 2017-12-15.
#> 2                                                                                                                                                                                                                               SpaceX new launch site currently under construction to help keep up with the Falcon 9 and Heavy manifests. Expected to be completed in late 2018. Initially will be limited to 12 flights per year, and only GTO launches.
#> 3                                                                                                                                                                                                                                                                                                                                            SpaceX west coast landing pad, has not yet been used. Expected to first be used during the Formosat-5 launch.
#> 4                                                                                                                                                                                                                    SpaceX east coast landing pad, where the historic first landing occurred. Originally used for early Atlas missiles and rockets from Lockheed Martin. Currently being expanded to add two smaller pads for Falcon Heavy RTLS missions.
#> 5                                                                                                                                                                                                                                                                                                      SpaceX original launch site, where all of the Falcon 1 launches occured. Abandoned as SpaceX decided against upgrading the pad to support Falcon 9.
#> 6 NASA historic launch pad that launched most of the Saturn V and Space Shuttle missions. Initially for Falcon Heavy launches, it is now launching all of SpaceX east coast missions due to the damage from the AMOS-6 anomaly. After SLC-40 repairs are complete, it will be upgraded to support Falcon Heavy, a process which will take about two months. In the future it will launch commercial crew missions and the Interplanetary Transport System.
#> 7                                                                                                                                                                                                                                                                                                   SpaceX original west coast launch pad for Falcon 1. Performed a static fire but was never used for a launch and abandoned due to scheduling conflicts.
#> 8                                                                                                                                                                                                                                                                                     SpaceX primary west coast launch pad for polar orbits and sun synchronous orbits, primarily used for Iridium. Also intended to be capable of launching Falcon Heavy.

## get details about launchpad "ksc_lc_39a"
get_SpaceX_launchpads("ksc_lc_39a")

#> $id
#> [1] "ksc_lc_39a"
#> 
#> $full_name
#> [1] "Kennedy Space Center Historic Launch Complex 39A"
#> 
#> $status
#> [1] "active"
#> 
#> $location
#> $location$name
#> [1] "Cape Canaveral"
#> 
#> $location$region
#> [1] "Florida"
#> 
#> $location$latitude
#> [1] 28.60806
#> 
#> $location$longitude
#> [1] -80.60396
#> 
#> 
#> $vehicles_launched
#> [1] "Falcon 9"     "Falcon Heavy"
#> 
#> $details
#> [1] "NASA historic launch pad that launched most of the Saturn V and Space Shuttle missions. Initially for Falcon Heavy launches, it is now launching all of SpaceX east coast missions due to the damage from the AMOS-6 anomaly. After SLC-40 repairs are complete, it will be upgraded to support Falcon Heavy, a process which will take about two months. In the future it will launch commercial crew missions and the Interplanetary Transport System."
#>