GET detailed info about SpaceX parts

GET detailed info about SpaceX parts

get_SpaceX_parts(type, name = NULL, ...)

Arguments

type
character: Type of parts. Choose between "caps" for capsule details or "cores" for core details.

name
character: Name of capsule or core.

...
: Query string to API. See more in details.

Value

data.frame with parsed data from SpaceX API.

Details

Query Strings for type = "caps"

Query Strings

Description

capsule_serial	Filter by capsule serial #

capsule_id

Filter by capsule id

status	Filter by capsule status

original_launch

Filter by original launch date

missions	Filter by flight missions

landings

Filter by # of landings

type	Filter by type of dragon capusle

Query Strings for type = "cores"

core_serial

Filter by core serial #

status	Filter by flight status

original_launch

Filter by original launch date

missions	Filter by flight missions

rtls_attempt

Filter by at least 1 rtls attempt

rtls_landings	Filter by total number of landings

asds_attempt

Filter by at least 1 asds attempt

asds_landings	Filter by total number of landings

water_landing

Filter by at least 1 water landing

Examples

## get details about all cores
get_SpaceX_parts(type = "cores")

#>    core_serial block    status          original_launch original_launch_unix
#> 1        B1052     5    active                     <NA>                   NA
#> 2        B1053     5    active                     <NA>                   NA
#> 3     Merlin1A    NA  expended 2006-03-24T22:30:00.000Z           1143239400
#> 4     Merlin2A    NA  expended 2007-03-21T01:10:00.000Z           1174439400
#> 5     Merlin1C    NA  expended 2008-08-02T03:34:00.000Z           1217648040
#> 6     Merlin2C    NA  expended 2008-09-28T23:15:00.000Z           1222643700
#> 7     Merlin3C    NA  expended 2009-07-13T03:35:00.000Z           1247456100
#> 8        B0003     1  expended 2010-06-04T18:45:00.000Z           1275677100
#> 9        B0004     1  expended 2010-12-08T15:43:00.000Z           1291822980
#> 10       B0005     1  expended 2012-05-22T07:44:00.000Z           1337672640
#> 11       B0006     1  expended 2012-10-08T00:35:00.000Z           1349656500
#> 12       B0007     1  expended 2013-03-01T19:10:00.000Z           1362165000
#> 13       B1003     1 destroyed 2013-09-29T16:00:00.000Z           1380470400
#> 14       B1004     1  expended 2013-12-03T22:41:00.000Z           1386110460
#> 15       B1005     1  expended 2014-01-06T18:06:00.000Z           1389031560
#> 16       B1006     1 destroyed 2014-04-18T19:25:00.000Z           1397849100
#> 17       B1007     1 destroyed 2014-07-14T15:15:00.000Z           1405350900
#> 18       B1008     1  expended 2014-08-05T08:00:00.000Z           1407225600
#> 19       B1011     1  expended 2014-09-07T05:00:00.000Z           1410066000
#> 20       B1010     1  expended 2014-09-21T05:52:00.000Z           1411278720
#> 21       B1012     1 destroyed 2015-01-10T09:47:00.000Z           1420883220
#> 22       B1013     1 destroyed 2015-02-11T23:03:00.000Z           1423695780
#> 23       B1014     1  expended 2015-03-02T03:50:00.000Z           1425268200
#> 24       B1015     1 destroyed 2015-04-14T20:10:00.000Z           1429042200
#> 25       B1016     1  expended 2015-04-27T23:03:00.000Z           1430175780
#> 26       B1018     1 destroyed 2015-06-28T14:21:00.000Z           1435501260
#> 27       B1019     1   retired 2015-12-22T01:29:00.000Z           1450747740
#> 28       B1017     1 destroyed 2016-01-17T15:42:00.000Z           1453045320
#> 29       B1020     1 destroyed 2016-03-04T23:35:00.000Z           1457134500
#> 30       B1021     2   retired 2016-04-08T20:43:00.000Z           1460148180
#> 31       B1022     2   retired 2016-05-06T05:21:00.000Z           1462512060
#> 32       B1023     2   retired 2016-05-27T21:39:00.000Z           1464385140
#> 33       B1024     2 destroyed 2016-06-15T14:29:00.000Z           1466000940
#> 34       B1025     2   retired 2016-07-18T04:45:00.000Z           1468817100
#> 35       B1026     2   retired 2016-08-14T05:26:00.000Z           1471152360
#> 36       B1028     3 destroyed 2016-09-01T13:07:00.000Z           1472735220
#> 37       B1029     3   retired 2017-01-14T17:54:00.000Z           1484416440
#> 38       B1031     3   retired 2017-02-19T14:39:00.000Z           1487515140
#> 39       B1030     3 destroyed 2017-03-16T06:00:00.000Z           1489644000
#> 40       B1032     3 destroyed 2017-05-01T11:15:00.000Z           1493637300
#> 41       B1034     3  expended 2017-05-15T23:21:00.000Z           1494890460
#> 42       B1035     3   retired 2017-06-03T21:07:00.000Z           1496524020
#> 43       B1036     3 destroyed 2017-06-25T13:25:00.000Z           1498397100
#> 44       B1037     3  expended 2017-07-05T23:35:00.000Z           1499297700
#> 45       B1039     4  expended 2017-08-14T16:31:00.000Z           1502728260
#> 46       B1038     3  expended 2017-08-24T18:50:00.000Z           1503600600
#> 47       B1040     4  expended 2017-09-07T13:50:00.000Z           1504792200
#> 48       B1041     4  expended 2017-10-09T12:37:00.000Z           1507552620
#> 49       B1042     4   retired 2017-10-30T19:34:00.000Z           1509392040
#> 50       B1043     4  expended 2017-11-11T23:00:00.000Z           1510441200
#> 51       B1033     3 destroyed 2018-02-06T20:45:00.000Z           1517949900
#> 52       B1044     4  expended 2018-03-06T05:33:00.000Z           1520314380
#> 53       B1045     4  expended 2018-04-18T22:51:00.000Z           1524091860
#> 54       B1046     5    active 2018-05-11T20:14:00.000Z           1526069640
#> 55       B1047     5    active 2018-07-22T05:50:00.000Z           1532238600
#> 56       B1048     5    active 2018-07-25T11:39:26.000Z           1532518766
#> 57       B1049     5    active 2018-09-10T04:45:00.000Z           1536554700
#> 58       B1051     5    active 2018-11-01T00:00:00.000Z           1541030400
#> 59       B1050     5    active 2018-12-15T00:00:00.000Z           1544832000
#>                                                  missions reuse_count
#> 1                                                    NULL           0
#> 2                                                    NULL           0
#> 3                                            FalconSat, 1           0
#> 4                                              DemoSat, 2           0
#> 5                                          Trailblazer, 3           0
#> 6                                               RatSat, 4           0
#> 7                                             RazakSat, 5           0
#> 8                                 Falcon 9 Test Flight, 6           0
#> 9                                               COTS 1, 7           0
#> 10                                              COTS 2, 8           0
#> 11                                               CRS-1, 9           0
#> 12                                              CRS-2, 10           0
#> 13                                           CASSIOPE, 11           0
#> 14                                              SES-8, 12           0
#> 15                                          Thaicom 6, 13           0
#> 16                                              CRS-3, 14           0
#> 17                                     OG-2 Mission 1, 15           0
#> 18                                          AsiaSat 8, 16           0
#> 19                                          AsiaSat 6, 17           0
#> 20                                              CRS-4, 18           0
#> 21                                              CRS-5, 19           0
#> 22                                             DSCOVR, 20           0
#> 23                           ABS-3A / Eutelsat 115W B, 21           0
#> 24                                              CRS-6, 22           0
#> 25                       TürkmenÄlem 52°E / MonacoSAT, 23           0
#> 26                                              CRS-7, 24           0
#> 27                                     OG-2 Mission 2, 25           0
#> 28                                            Jason 3, 26           0
#> 29                                              SES-9, 27           0
#> 30                                  CRS-8, SES-10, 28, 38           1
#> 31                                           JCSAT-2B, 29           0
#> 32            Thaicom 8, Falcon Heavy Test Flight, 30, 55           1
#> 33                           ABS-2A / Eutelsat 117W B, 31           0
#> 34                CRS-9, Falcon Heavy Test Flight, 32, 55           1
#> 35                                           JCSAT-16, 33           0
#> 36                                             Amos-6, 34           0
#> 37          Iridium NEXT Mission 1, BulgariaSat-1, 35, 42           1
#> 38                  CRS-10, SES-11 / Echostar 105, 36, 49           1
#> 39                                        EchoStar 23, 37           0
#> 40                     NROL-76, SES-16 / GovSat-1, 39, 54           1
#> 41                                      Inmarsat-5 F4, 40           0
#> 42                                 CRS-11, CRS-13, 41, 51           1
#> 43 Iridium NEXT Mission 2, Iridium NEXT Mission 4, 43, 52           1
#> 44                                       Intelsat 35e, 44           0
#> 45                                 CRS-12, CRS-14, 45, 59           1
#> 46                FormoSat-5, Paz / Starlink Demo, 46, 56           1
#> 47                     Boeing X-37B OTV-5, SES-12, 47, 63           1
#> 48 Iridium NEXT Mission 3, Iridium NEXT Mission 5, 48, 58           1
#> 49                                        KoreaSat 5A, 50           0
#> 50                   ZUMA, Iridium NEXT Mission 6, 53, 62           1
#> 51                           Falcon Heavy Test Flight, 55           0
#> 52                                     Hispasat 30W-6, 57           0
#> 53                                   TESS, CRS-15, 60, 64           1
#> 54                     Bangabandhu-1, Merah Putih, 61, 67           1
#> 55                         Telstar 19V, Es’hail 2, 65, 70           1
#> 56              Iridium NEXT Mission 7, SAOCOM 1A, 66, 69           1
#> 57                                        Telstar 18V, 68           0
#> 58                                                   NULL           0
#> 59                                                   NULL           0
#>    rtls_attempts rtls_landings asds_attempts asds_landings water_landing
#> 1              0             0             0             0         FALSE
#> 2              0             0             0             0         FALSE
#> 3              0             0             0             0         FALSE
#> 4              0             0             0             0         FALSE
#> 5              0             0             0             0         FALSE
#> 6              0             0             0             0         FALSE
#> 7              0             0             0             0         FALSE
#> 8              0             0             0             0         FALSE
#> 9              0             0             0             0         FALSE
#> 10             0             0             0             0         FALSE
#> 11             0             0             0             0         FALSE
#> 12             0             0             0             0         FALSE
#> 13             0             0             0             0          TRUE
#> 14             0             0             0             0         FALSE
#> 15             0             0             0             0         FALSE
#> 16             0             0             0             0          TRUE
#> 17             0             0             0             0          TRUE
#> 18             0             0             0             0         FALSE
#> 19             0             0             0             0         FALSE
#> 20             0             0             0             0          TRUE
#> 21             0             0             1             0         FALSE
#> 22             0             0             0             0          TRUE
#> 23             0             0             0             0         FALSE
#> 24             0             0             1             0         FALSE
#> 25             0             0             0             0         FALSE
#> 26             0             0             0             0         FALSE
#> 27             1             1             0             0         FALSE
#> 28             0             0             1             0         FALSE
#> 29             0             0             1             0         FALSE
#> 30             0             0             2             2         FALSE
#> 31             0             0             1             1         FALSE
#> 32             1             1             1             1         FALSE
#> 33             0             0             1             0         FALSE
#> 34             2             2             0             0         FALSE
#> 35             0             0             1             1         FALSE
#> 36             0             0             0             0         FALSE
#> 37             0             0             2             2         FALSE
#> 38             1             1             1             1         FALSE
#> 39             0             0             0             0         FALSE
#> 40             1             1             0             0          TRUE
#> 41             0             0             0             0         FALSE
#> 42             2             2             0             0         FALSE
#> 43             0             0             1             1          TRUE
#> 44             0             0             0             0         FALSE
#> 45             1             1             0             0         FALSE
#> 46             0             0             1             1         FALSE
#> 47             1             1             0             0         FALSE
#> 48             0             0             1             1         FALSE
#> 49             0             0             1             1         FALSE
#> 50             1             1             0             0         FALSE
#> 51             0             0             1             0         FALSE
#> 52             0             0             0             0         FALSE
#> 53             0             0             1             1         FALSE
#> 54             0             0             2             2         FALSE
#> 55             0             0             2             2         FALSE
#> 56             1             1             1             1         FALSE
#> 57             0             0             1             1         FALSE
#> 58             0             0             0             0         FALSE
#> 59             0             0             0             0         FALSE
#>                                                                                                                                                                                       details
#> 1                                                                                                                                                                  On test stand at McGregor.
#> 2                                                                                                                                                                        Testing at McGregor.
#> 3                                                                                                                                  Engine failure at T+33 seconds resulted in loss of vehicle
#> 4  Successful first-stage burn and transition to second stage, maximal altitude 289 km. Harmonic oscillation at T+5 minutes Premature engine shutdown at T+7 min 30 s. Failed to reach orbit.
#> 5                                                                                                                       Residual stage-1 thrust led to collision between stage 1 and stage 2.
#> 6                                                                      Initially scheduled for 23–25 Sep, carried dummy payload – mass simulator, 165 kg (originally intended to be RazakSAT.
#> 7                                                                                                                                                                                        <NA>
#> 8                                                                                                                       Core expended on flight, no recovery effort. First flight of Falcon 9
#> 9                                                                                                                                                                      First flight of Dragon
#> 10                                                                                                                                                                                       <NA>
#> 11                                                                                                                               Suffered engine out at T+1:19 but primary mission successful
#> 12                                                                                                                                                                                       <NA>
#> 13                                                                                                                 First flight of Falcon 9 v1.1 upgrade, first Spacex flight from Vandenberg
#> 14                                                                                                                                                                                       <NA>
#> 15                                                                                                                                                                                       <NA>
#> 16                                                                                                                                                     Broke up after sucessful water landing
#> 17                                                                                                                                                     Broke up after sucessful water landing
#> 18                                                                                                                                                                                       <NA>
#> 19                                                                                                                                                                                       <NA>
#> 20                                                                                                                                                     Broke up after sucessful water landing
#> 21                                                                                                                      Destroyed on impact with droneship, grid fin hydraulic fluid depleted
#> 22                                                                                                                                                     Broke up after sucessful water landing
#> 23                                                                                                                                                                                       <NA>
#> 24                                                                                                      Destroyed on impact with droneship due to excess lateral velocity - throttle stiction
#> 25                                                                                                                                                                                       <NA>
#> 26                                                                                                   RUD at T + 139s - second stage overpressure event due to defective COPV strut heim joint
#> 27                                                                                                                                         First recovered stage, first return to launch site
#> 28                                                                                                                              Destroyed after drone ship landing due to landing leg failure
#> 29                                                                                                       Destroyed on impact with drone ship due to excess velocity, possible fuel exhaustion
#> 30                                         First drone ship landing, first orbital booster to be reflown, first orbital booster to be landed twice, will be given as a gift to Cape Canaveral
#> 31                                                                                        First recovered booster from a GTO mission, in storage, and due to be retired after ten test cycles
#> 32                                                                                                                         Converted to side booster for Falcon Heavy, retired after FH Demo.
#> 33                                                                                                                                                                  RUD on deck of drone ship
#> 34                                                                                                                         Converted to side booster for Falcon Heavy, retired after FH Demo.
#> 35                                                                                                                                                                       Permanently scrapped
#> 36                                                                                                                                          Destroyed on launchpad during static fire anomaly
#> 37                                                                                      Second booster to be reflown, Completed sucessful launch of BulgariaSat-1 and ASDS landing on 6/23/17
#> 38                                                                                                                                      Daytime RTLS landing, helped launch CRS-10 to the ISS
#> 39                                                                                                                                                                                       <NA>
#> 40                                                                                                                       Daytime RTLS landing, helped launch first SpaceX mission for the NRO
#> 41                                                                                                                                                                                       <NA>
#> 42                                                                                                                                      Daytime RTLS landing, helped launch CRS-11 to the ISS
#> 43                                                                                                                                                                  Destroyed by waterlanding
#> 44                                                                                                                                        Core will be expended on flight, no recovery effort
#> 45                                                                                                                                                                  Flew expendable on CRS-14
#> 46                                                                                                               Helped launch FormoSat-5. Will now fly expendable on Paz & Microsat-2a, -2b.
#> 47                                                                                                                                                                                       <NA>
#> 48                                                                                                                                              Will fly expendable on Iridium NEXT Mission 5
#> 49                                                                                                                   The Octaweb caught on fire right after Landing due to an RP-1 fuel leak.
#> 50                                                                                                                                                                                       <NA>
#> 51                                                                                                                                                    Failed landing after Falcon Heavy Demo.
#> 52                                                                                                      Core expended on Hispasat flight. Water landing attempted, but the result is unknown.
#> 53                                                                                                                                                                                       <NA>
#> 54                                                                                                                                                                        First block V core.
#> 55                                                                                                                                                                                       <NA>
#> 56                                                                                                                                                                                       <NA>
#> 57                                                                                                                                                                                       <NA>
#> 58                                                                                                                                                                               At McGregor.
#> 59                                                                                                                                                                En-route to Cape Canaveral.

## get details about all capsules
get_SpaceX_parts(type = "caps")

#>    capsule_serial capsule_id    status          original_launch
#> 1            C201    dragon2    active                     <NA>
#> 2            C202    dragon2    active                     <NA>
#> 3            C203    dragon2    active                     <NA>
#> 4            C204    dragon2    active                     <NA>
#> 5            C205    dragon2    active                     <NA>
#> 6            C101    dragon1   retired 2010-12-08T15:43:00.000Z
#> 7            C102    dragon1   retired 2012-05-02T07:44:00.000Z
#> 8            C103    dragon1   unknown 2012-10-08T00:35:00.000Z
#> 9            C104    dragon1   unknown 2013-03-01T19:10:00.000Z
#> 10           C105    dragon1   unknown 2014-04-18T19:25:00.000Z
#> 11           C106    dragon1    active 2014-09-21T05:52:00.000Z
#> 12           C107    dragon1   unknown 2015-01-10T09:47:00.000Z
#> 13           C108    dragon1    active 2015-04-14T20:10:00.000Z
#> 14           C109    dragon1 destroyed 2015-06-28T14:21:00.000Z
#> 15           C110    dragon1    active 2016-04-08T20:43:00.000Z
#> 16           C111    dragon1    active 2016-07-18T04:45:00.000Z
#> 17           C112    dragon1    active 2017-02-19T14:39:00.000Z
#> 18           C113    dragon1    active 2017-08-14T16:31:00.000Z
#>    original_launch_unix              missions landings       type
#> 1                    NA                  NULL        0 Dragon 2.0
#> 2                    NA                  NULL        0 Dragon 2.0
#> 3                    NA                  NULL        0 Dragon 2.0
#> 4                    NA                  NULL        0 Dragon 2.0
#> 5                    NA                  NULL        0 Dragon 2.0
#> 6            1291822980             COTS 1, 7        0 Dragon 1.0
#> 7            1335944640             COTS 2, 8        1 Dragon 1.0
#> 8            1349656500              CRS-1, 9        1 Dragon 1.0
#> 9            1362165000             CRS-2, 10        1 Dragon 1.0
#> 10           1397849100             CRS-3, 14        1 Dragon 1.1
#> 11           1411278720 CRS-4, CRS-11, 18, 41        2 Dragon 1.1
#> 12           1420883220             CRS-5, 19        1 Dragon 1.1
#> 13           1429042200 CRS-6, CRS-13, 22, 51        2 Dragon 1.1
#> 14           1435501260             CRS-7, 24        0 Dragon 1.1
#> 15           1460148180 CRS-8, CRS-14, 28, 59        2 Dragon 1.1
#> 16           1468817100 CRS-9, CRS-15, 32, 64        2 Dragon 1.1
#> 17           1487515140            CRS-10, 36        1 Dragon 1.1
#> 18           1502728260            CRS-12, 45        1 Dragon 1.1
#>                                                                                                        details
#> 1  Pressure vessel used for Dragon 2 structural testing. Rumored to be repurposed for first Red Dragon Mission
#> 2                           Capsule used to qualify Dragon 2's environmental control and life support systems.
#> 3                                                        Rumored to be used for Inflight Abort Test after DM-1
#> 4                                                                    Currently in construction for use in DM-2
#> 5                               In construction for use in first mission in contract under the CCtCap contract
#> 6                                                                         Reentered after three weeks in orbit
#> 7                                                                                      First Dragon spacecraft
#> 8                                                       First of twenty missions flown under the CRS1 contract
#> 9                                                                                                         <NA>
#> 10                                                                                   First Dragon v1.1 capsule
#> 11                                                                           First Dragon capsule to be reused
#> 12                                                                                                        <NA>
#> 13                                                                          Second Dragon capsule to be reused
#> 14                                                                 Destroyed on impact after F9 launch failure
#> 15                                                                                                        <NA>
#> 16                                                                                                        <NA>
#> 17                                                                                                        <NA>
#> 18                                                                        The last newly manufactured Dragon 1
#>    reuse_count
#> 1            0
#> 2            0
#> 3            0
#> 4            0
#> 5            0
#> 6            0
#> 7            0
#> 8            0
#> 9            0
#> 10           0
#> 11           1
#> 12           0
#> 13           1
#> 14           0
#> 15           1
#> 16           1
#> 17           0
#> 18           0

## get details about all active capsules
get_SpaceX_parts(type = "caps", status = "active")

#>    capsule_serial capsule_id status          original_launch
#> 1            C201    dragon2 active                     <NA>
#> 2            C202    dragon2 active                     <NA>
#> 3            C203    dragon2 active                     <NA>
#> 4            C204    dragon2 active                     <NA>
#> 5            C205    dragon2 active                     <NA>
#> 6            C106    dragon1 active 2014-09-21T05:52:00.000Z
#> 7            C108    dragon1 active 2015-04-14T20:10:00.000Z
#> 8            C110    dragon1 active 2016-04-08T20:43:00.000Z
#> 9            C111    dragon1 active 2016-07-18T04:45:00.000Z
#> 10           C112    dragon1 active 2017-02-19T14:39:00.000Z
#> 11           C113    dragon1 active 2017-08-14T16:31:00.000Z
#>    original_launch_unix              missions landings       type
#> 1                    NA                  NULL        0 Dragon 2.0
#> 2                    NA                  NULL        0 Dragon 2.0
#> 3                    NA                  NULL        0 Dragon 2.0
#> 4                    NA                  NULL        0 Dragon 2.0
#> 5                    NA                  NULL        0 Dragon 2.0
#> 6            1411278720 CRS-4, CRS-11, 18, 41        2 Dragon 1.1
#> 7            1429042200 CRS-6, CRS-13, 22, 51        2 Dragon 1.1
#> 8            1460148180 CRS-8, CRS-14, 28, 59        2 Dragon 1.1
#> 9            1468817100 CRS-9, CRS-15, 32, 64        2 Dragon 1.1
#> 10           1487515140            CRS-10, 36        1 Dragon 1.1
#> 11           1502728260            CRS-12, 45        1 Dragon 1.1
#>                                                                                                        details
#> 1  Pressure vessel used for Dragon 2 structural testing. Rumored to be repurposed for first Red Dragon Mission
#> 2                           Capsule used to qualify Dragon 2's environmental control and life support systems.
#> 3                                                        Rumored to be used for Inflight Abort Test after DM-1
#> 4                                                                    Currently in construction for use in DM-2
#> 5                               In construction for use in first mission in contract under the CCtCap contract
#> 6                                                                            First Dragon capsule to be reused
#> 7                                                                           Second Dragon capsule to be reused
#> 8                                                                                                         <NA>
#> 9                                                                                                         <NA>
#> 10                                                                                                        <NA>
#> 11                                                                        The last newly manufactured Dragon 1
#>    reuse_count
#> 1            0
#> 2            0
#> 3            0
#> 4            0
#> 5            0
#> 6            1
#> 7            1
#> 8            1
#> 9            1
#> 10           0
#> 11           0

## get details about core B1041
get_SpaceX_parts(type = "cores", name = "B1041")

#> $core_serial
#> [1] "B1041"
#> 
#> $block
#> [1] 4
#> 
#> $status
#> [1] "expended"
#> 
#> $original_launch
#> [1] "2017-10-09T12:37:00.000Z"
#> 
#> $original_launch_unix
#> [1] 1507552620
#> 
#> $missions
#>                     name flight
#> 1 Iridium NEXT Mission 3     48
#> 2 Iridium NEXT Mission 5     58
#> 
#> $reuse_count
#> [1] 1
#> 
#> $rtls_attempts
#> [1] 0
#> 
#> $rtls_landings
#> [1] 0
#> 
#> $asds_attempts
#> [1] 1
#> 
#> $asds_landings
#> [1] 1
#> 
#> $water_landing
#> [1] FALSE
#> 
#> $details
#> [1] "Will fly expendable on Iridium NEXT Mission 5"
#>