GET info about gives orbital parameters about the current position of Elon's Tesla Roadster that was launched on the first Falcon Heavy demo flight.

GET info about gives orbital parameters about the current position of Elon's Tesla Roadster that was launched on the first Falcon Heavy demo flight.

get_SpaceX_roadster(...)

Arguments

...
: Query string to API. See more in details.

Value

data.frame with parsed data from SpaceX API.

Examples

## get details about the Tesla roadster position
get_SpaceX_roadster()

#> $name
#> [1] "Elon Musk's Tesla Roadster"
#> 
#> $launch_date_utc
#> [1] "2018-02-06T20:45:00.000Z"
#> 
#> $launch_date_unix
#> [1] 1517949900
#> 
#> $launch_mass_kg
#> [1] 1350
#> 
#> $launch_mass_lbs
#> [1] 2976
#> 
#> $norad_id
#> [1] 43205
#> 
#> $epoch_jd
#> [1] 2459005
#> 
#> $orbit_type
#> [1] "heliocentric"
#> 
#> $apoapsis_au
#> [1] 1.663951
#> 
#> $periapsis_au
#> [1] 0.9859809
#> 
#> $semi_major_axis_au
#> [1] 190.5456
#> 
#> $eccentricity
#> [1] 0.2558443
#> 
#> $inclination
#> [1] 1.077498
#> 
#> $longitude
#> [1] 317.084
#> 
#> $periapsis_arg
#> [1] 177.5233
#> 
#> $period_days
#> [1] 557.0644
#> 
#> $speed_kph
#> [1] 71917.13
#> 
#> $speed_mph
#> [1] 44687.22
#> 
#> $earth_distance_km
#> [1] 234079045
#> 
#> $earth_distance_mi
#> [1] 145449930
#> 
#> $mars_distance_km
#> [1] 96381008
#> 
#> $mars_distance_mi
#> [1] 59888363
#> 
#> $flickr_images
#> [1] "https://farm5.staticflickr.com/4615/40143096241_11128929df_b.jpg"
#> [2] "https://farm5.staticflickr.com/4702/40110298232_91b32d0cc0_b.jpg"
#> [3] "https://farm5.staticflickr.com/4676/40110297852_5e794b3258_b.jpg"
#> [4] "https://farm5.staticflickr.com/4745/40110304192_6e3e9a7a1b_b.jpg"
#> 
#> $wikipedia
#> [1] "https://en.wikipedia.org/wiki/Elon_Musk%27s_Tesla_Roadster"
#> 
#> $video
#> [1] "https://youtu.be/wbSwFU6tY1c"
#> 
#> $details
#> [1] "Elon Musk's Tesla Roadster is an electric sports car that served as the dummy payload for the February 2018 Falcon Heavy test flight and is now an artificial satellite of the Sun. Starman, a mannequin dressed in a spacesuit, occupies the driver's seat. The car and rocket are products of Tesla and SpaceX. This 2008-model Roadster was previously used by Musk for commuting, and is the only consumer car sent into space."
#>