GET info about SpaceX rockets

GET info about SpaceX rockets

get_SpaceX_rockets(name = NULL)

Arguments

name
character: Name of the rocket. Choose between "falcon1", "falcon9" or "falconheavy". By default all three rocket data are returned.

Value

data.frame with parsed data from SpaceX API.

Examples

## get details about all rockets
get_SpaceX_rockets()

#>            id         name   type active stages boosters cost_per_launch
#> 1     falcon1     Falcon 1 rocket  FALSE      2        0        67000000
#> 2     falcon9     Falcon 9 rocket   TRUE      2        0        61200000
#> 3 falconheavy Falcon Heavy rocket  FALSE      2        2        90000000
#>   success_rate_pct first_flight                          country company
#> 1               40   2006-03-24 Republic of the Marshall Islands  SpaceX
#> 2               94   2010-06-04                    United States  SpaceX
#> 3                0          TBD                    United States  SpaceX
#>   height.meters height.feet diameter.meters diameter.feet mass.kg mass.lb
#> 1         22.25        73.0            1.68           5.5   30146   66460
#> 2         70.00       229.6            3.70          12.0  549054 1207920
#> 3         70.00       229.6           12.20          39.9 1420788 3125735
#>                                                                                                                                         payload_weights
#> 1                                                                                                                        leo, Low Earth Orbit, 450, 992
#> 2                                     leo, gto, mars, Low Earth Orbit, Geosynchronous Transfer Orbit, Mars Orbit, 22800, 8300, 4020, 50265, 18300, 8860
#> 3 leo, gto, mars, pluto, Low Earth Orbit, Geosynchronous Transfer Orbit, Mars Orbit, Pluto Orbit, 63800, 26700, 16800, 3500, 140660, 58860, 37040, 7720
#>   first_stage.reusable first_stage.engines first_stage.fuel_amount_tons
#> 1                FALSE                   1                         44.3
#> 2                 TRUE                   9                        385.0
#> 3                 TRUE                  27                       1155.0
#>   first_stage.burn_time_sec first_stage.thrust_sea_level.kN
#> 1                       169                             420
#> 2                       162                            7607
#> 3                       162                           22819
#>   first_stage.thrust_sea_level.lbf first_stage.thrust_vacuum.kN
#> 1                            94000                          480
#> 2                          1710000                         8227
#> 3                          5130000                        24681
#>   first_stage.thrust_vacuum.lbf first_stage.cores second_stage.engines
#> 1                        110000                NA                    1
#> 2                       1849500                NA                    1
#> 3                       5548500                 3                    1
#>   second_stage.fuel_amount_tons second_stage.burn_time_sec
#> 1                          3.38                        378
#> 2                         90.00                        397
#> 3                            NA                        397
#>   second_stage.thrust.kN second_stage.thrust.lbf second_stage.payloads.option_1
#> 1                     31                    7000              composite fairing
#> 2                    934                  210000                         dragon
#> 3                    934                  210000                         dragon
#>   second_stage.payloads.composite_fairing.height.meters
#> 1                                                   3.5
#> 2                                                  13.1
#> 3                                                  13.1
#>   second_stage.payloads.composite_fairing.height.feet
#> 1                                                11.5
#> 2                                                43.0
#> 3                                                43.0
#>   second_stage.payloads.composite_fairing.diameter.meters
#> 1                                                     1.5
#> 2                                                     5.2
#> 3                                                     5.2
#>   second_stage.payloads.composite_fairing.diameter.feet
#> 1                                                   4.9
#> 2                                                  17.1
#> 3                                                  17.1
#>   second_stage.payloads.option_2 engines.number engines.type engines.version
#> 1                           <NA>              1       merlin              1C
#> 2              composite fairing              9       merlin             1D+
#> 3              composite fairing             27       merlin             1D+
#>   engines.layout engines.engine_loss_max    engines.propellant_1
#> 1         single                       0           liquid oxygen
#> 2        octaweb                       2           liquid oxygen
#> 3        octaweb                       6 subcooled liquid oxygen
#>   engines.propellant_2 engines.thrust_sea_level.kN engines.thrust_sea_level.lbf
#> 1        RP-1 kerosene                         420                        94000
#> 2        RP-1 kerosene                         845                       190000
#> 3        RP-1 kerosene                         845                       190000
#>   engines.thrust_vacuum.kN engines.thrust_vacuum.lbf engines.thrust_to_weight
#> 1                      480                    110000                     96.0
#> 2                      914                    205500                    180.1
#> 3                      914                    205500                    180.1
#>   landing_legs.number landing_legs.material
#> 1                   0                  <NA>
#> 2                   4          carbon fiber
#> 3                  12          carbon fiber
#>                                                                                                                                                                                                                                                                                       description
#> 1                                                 The Falcon 1 was an expendable launch system privately developed and manufactured by SpaceX during 2006-2009. On 28 September 2008, Falcon 1 became the first privately-developed liquid-fuel launch vehicle to go into orbit around the Earth.
#> 2                                                                                                                                      Falcon 9 is a two-stage rocket designed and manufactured by SpaceX for the reliable and safe transport of satellites and the Dragon spacecraft into orbit.
#> 3 With the ability to lift into orbit over 54 metric tons (119,000 lb)--a mass equivalent to a 737 jetliner loaded with passengers, crew, luggage and fuel--Falcon Heavy can lift more than twice the payload of the next closest operational vehicle, the Delta IV Heavy, at one-third the cost.

## get details about "falcon1" rocket
get_SpaceX_rockets("falcon1")

#> $id
#> [1] "falcon1"
#> 
#> $name
#> [1] "Falcon 1"
#> 
#> $type
#> [1] "rocket"
#> 
#> $active
#> [1] FALSE
#> 
#> $stages
#> [1] 2
#> 
#> $boosters
#> [1] 0
#> 
#> $cost_per_launch
#> [1] 67000000
#> 
#> $success_rate_pct
#> [1] 40
#> 
#> $first_flight
#> [1] "2006-03-24"
#> 
#> $country
#> [1] "Republic of the Marshall Islands"
#> 
#> $company
#> [1] "SpaceX"
#> 
#> $height
#> $height$meters
#> [1] 22.25
#> 
#> $height$feet
#> [1] 73
#> 
#> 
#> $diameter
#> $diameter$meters
#> [1] 1.68
#> 
#> $diameter$feet
#> [1] 5.5
#> 
#> 
#> $mass
#> $mass$kg
#> [1] 30146
#> 
#> $mass$lb
#> [1] 66460
#> 
#> 
#> $payload_weights
#>    id            name  kg  lb
#> 1 leo Low Earth Orbit 450 992
#> 
#> $first_stage
#> $first_stage$reusable
#> [1] FALSE
#> 
#> $first_stage$engines
#> [1] 1
#> 
#> $first_stage$fuel_amount_tons
#> [1] 44.3
#> 
#> $first_stage$burn_time_sec
#> [1] 169
#> 
#> $first_stage$thrust_sea_level
#> $first_stage$thrust_sea_level$kN
#> [1] 420
#> 
#> $first_stage$thrust_sea_level$lbf
#> [1] 94000
#> 
#> 
#> $first_stage$thrust_vacuum
#> $first_stage$thrust_vacuum$kN
#> [1] 480
#> 
#> $first_stage$thrust_vacuum$lbf
#> [1] 110000
#> 
#> 
#> 
#> $second_stage
#> $second_stage$engines
#> [1] 1
#> 
#> $second_stage$fuel_amount_tons
#> [1] 3.38
#> 
#> $second_stage$burn_time_sec
#> [1] 378
#> 
#> $second_stage$thrust
#> $second_stage$thrust$kN
#> [1] 31
#> 
#> $second_stage$thrust$lbf
#> [1] 7000
#> 
#> 
#> $second_stage$payloads
#> $second_stage$payloads$option_1
#> [1] "composite fairing"
#> 
#> $second_stage$payloads$composite_fairing
#> $second_stage$payloads$composite_fairing$height
#> $second_stage$payloads$composite_fairing$height$meters
#> [1] 3.5
#> 
#> $second_stage$payloads$composite_fairing$height$feet
#> [1] 11.5
#> 
#> 
#> $second_stage$payloads$composite_fairing$diameter
#> $second_stage$payloads$composite_fairing$diameter$meters
#> [1] 1.5
#> 
#> $second_stage$payloads$composite_fairing$diameter$feet
#> [1] 4.9
#> 
#> 
#> 
#> 
#> 
#> $engines
#> $engines$number
#> [1] 1
#> 
#> $engines$type
#> [1] "merlin"
#> 
#> $engines$version
#> [1] "1C"
#> 
#> $engines$layout
#> [1] "single"
#> 
#> $engines$engine_loss_max
#> [1] 0
#> 
#> $engines$propellant_1
#> [1] "liquid oxygen"
#> 
#> $engines$propellant_2
#> [1] "RP-1 kerosene"
#> 
#> $engines$thrust_sea_level
#> $engines$thrust_sea_level$kN
#> [1] 420
#> 
#> $engines$thrust_sea_level$lbf
#> [1] 94000
#> 
#> 
#> $engines$thrust_vacuum
#> $engines$thrust_vacuum$kN
#> [1] 480
#> 
#> $engines$thrust_vacuum$lbf
#> [1] 110000
#> 
#> 
#> $engines$thrust_to_weight
#> [1] 96
#> 
#> 
#> $landing_legs
#> $landing_legs$number
#> [1] 0
#> 
#> $landing_legs$material
#> NULL
#> 
#> 
#> $description
#> [1] "The Falcon 1 was an expendable launch system privately developed and manufactured by SpaceX during 2006-2009. On 28 September 2008, Falcon 1 became the first privately-developed liquid-fuel launch vehicle to go into orbit around the Earth."
#>