Get location from one EcoCounter installation

Get location from one EcoCounter installation

get_EcoCounter_data(EcoCounterId, from = NULL, to = NULL, step = 4)

Arguments

EcoCounterId
character or integer or vector of character or integer: Id of the EcoCounter installation

from
character: Start date (YYYYMMDD). If no date is submitted, the day before yesterday is automatically set.

to
character: End date (YYYYMMDD). If no date is submitted, yesterday is automatically set. Note: When date is the date of today, usually no data are available.

step
integer: Get data : 2 (15 min), 3 (hourly), 4 (daily), 5 (weekly), 6 (monthly), 7 (yearly). With default: 4.

Value

data.frame with parsed data from EcoCounter API

Function version

0.0.1

Examples

EcoCounterId <- 100037011
get_EcoCounter_data(EcoCounterId)

#>      date      id      
#> [1,] Numeric,0 factor,1

## set specific date and request weekly data (step = 5)
get_EcoCounter_data(EcoCounterId, from = "20171101", to = "20171201", step = 5)

#>         date comptage        id
#> 1 2017-11-01      284 100037011
#> 2 2017-11-06      361 100037011
#> 3 2017-11-13      284 100037011
#> 4 2017-11-20      236 100037011
#> 5 2017-11-27      141 100037011