GET info about SpaceX launchpads

GET info about SpaceX launchpads

get_SpaceX_launchpads(name = NULL)

Arguments

name
character: Name of lauchnpads. Choose "ksc_lc_39a" or NULL. By default all three rocket data are returned.

Value

data.frame with parsed data from SpaceX API.

Examples

## get details about all launchpads
get_SpaceX_launchpads()

#>   padid              id
#> 1     6     vafb_slc_4e
#> 2     1 kwajalein_atoll
#> 3     2    ccafs_slc_40
#> 4     8            stls
#> 5     4      ksc_lc_39a
#> 6     5     vafb_slc_3w
#>                                                  full_name             status
#> 1        Vandenberg Air Force Base Space Launch Complex 4E             active
#> 2                            Kwajalein Atoll Omelek Island            retired
#> 3 Cape Canaveral Air Force Station Space Launch Complex 40             active
#> 4                           SpaceX South Texas Launch Site under construction
#> 5         Kennedy Space Center Historic Launch Complex 39A             active
#> 6        Vandenberg Air Force Base Space Launch Complex 3W            retired
#>               location.name  location.region location.latitude
#> 1 Vandenberg Air Force Base       California         34.632093
#> 2             Omelek Island Marshall Islands          9.047721
#> 3            Cape Canaveral          Florida         28.561857
#> 4        Boca Chica Village            Texas         25.997264
#> 5            Cape Canaveral          Florida         28.608058
#> 6 Vandenberg Air Force Base       California         34.644090
#>   location.longitude      vehicles_launched attempted_launches
#> 1         -120.61083               Falcon 9                 12
#> 2          167.74313               Falcon 1                  5
#> 3          -80.57737               Falcon 9                 38
#> 4          -97.15608               Falcon 9                  0
#> 5          -80.60396 Falcon 9, Falcon Heavy                 15
#> 6         -120.59314               Falcon 1                  0
#>   successful_launches
#> 1                  12
#> 2                   2
#> 3                  36
#> 4                   0
#> 5                  15
#> 6                   0
#>                                                                                wikipedia
#> 1                    https://en.wikipedia.org/wiki/Vandenberg_AFB_Space_Launch_Complex_4
#> 2                                            https://en.wikipedia.org/wiki/Omelek_Island
#> 3 https://en.wikipedia.org/wiki/Cape_Canaveral_Air_Force_Station_Space_Launch_Complex_40
#> 4                           https://en.wikipedia.org/wiki/SpaceX_South_Texas_Launch_Site
#> 5    https://en.wikipedia.org/wiki/Kennedy_Space_Center_Launch_Complex_39#Launch_Pad_39A
#> 6                    https://en.wikipedia.org/wiki/Vandenberg_AFB_Space_Launch_Complex_3
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                    details
#> 1                                                                                                                                                                                                                                                                                     SpaceX primary west coast launch pad for polar orbits and sun synchronous orbits, primarily used for Iridium. Also intended to be capable of launching Falcon Heavy.
#> 2                                                                                                                                                                                                                                                                                                      SpaceX original launch site, where all of the Falcon 1 launches occured. Abandoned as SpaceX decided against upgrading the pad to support Falcon 9.
#> 3                                                                                                                                                                                                                                             SpaceX primary Falcon 9 launch pad, where all east coast Falcon 9s launched prior to the AMOS-6 anomaly. Initially used to launch Titan rockets for Lockheed Martin. Back online since CRS-13 on 2017-12-15.
#> 4                                                                                                                                                                                                                               SpaceX new launch site currently under construction to help keep up with the Falcon 9 and Heavy manifests. Expected to be completed in late 2018. Initially will be limited to 12 flights per year, and only GTO launches.
#> 5 NASA historic launch pad that launched most of the Saturn V and Space Shuttle missions. Initially for Falcon Heavy launches, it is now launching all of SpaceX east coast missions due to the damage from the AMOS-6 anomaly. After SLC-40 repairs are complete, it will be upgraded to support Falcon Heavy, a process which will take about two months. In the future it will launch commercial crew missions and the Interplanetary Transport System.
#> 6                                                                                                                                                                                                                                                                                                   SpaceX original west coast launch pad for Falcon 1. Performed a static fire but was never used for a launch and abandoned due to scheduling conflicts.

## get details about launchpad "ksc_lc_39a"
get_SpaceX_launchpads("ksc_lc_39a")

#> $padid
#> [1] 4
#> 
#> $id
#> [1] "ksc_lc_39a"
#> 
#> $full_name
#> [1] "Kennedy Space Center Historic Launch Complex 39A"
#> 
#> $status
#> [1] "active"
#> 
#> $location
#> $location$name
#> [1] "Cape Canaveral"
#> 
#> $location$region
#> [1] "Florida"
#> 
#> $location$latitude
#> [1] 28.60806
#> 
#> $location$longitude
#> [1] -80.60396
#> 
#> 
#> $vehicles_launched
#> [1] "Falcon 9"     "Falcon Heavy"
#> 
#> $attempted_launches
#> [1] 15
#> 
#> $successful_launches
#> [1] 15
#> 
#> $wikipedia
#> [1] "https://en.wikipedia.org/wiki/Kennedy_Space_Center_Launch_Complex_39#Launch_Pad_39A"
#> 
#> $details
#> [1] "NASA historic launch pad that launched most of the Saturn V and Space Shuttle missions. Initially for Falcon Heavy launches, it is now launching all of SpaceX east coast missions due to the damage from the AMOS-6 anomaly. After SLC-40 repairs are complete, it will be upgraded to support Falcon Heavy, a process which will take about two months. In the future it will launch commercial crew missions and the Interplanetary Transport System."
#>