GET info about SpaceX launchpads

GET info about SpaceX launchpads

get_SpaceX_launchpads(...)

Arguments

...
: Query string to API. See more in details.

Value

data.frame with parsed data from SpaceX API.

Examples

## get details about all launchpads
get_SpaceX_launchpads()

#>   id            name             status             location.name
#> 1  5     VAFB SLC 3W            retired Vandenberg Air Force Base
#> 2  2    CCAFS SLC 40             active            Cape Canaveral
#> 3  8            STLS under construction        Boca Chica Village
#> 4  1 Kwajalein Atoll            retired             Omelek Island
#> 5  6     VAFB SLC 4E             active Vandenberg Air Force Base
#> 6  4      KSC LC 39A             active            Cape Canaveral
#>    location.region location.latitude location.longitude      vehicles_launched
#> 1       California         34.644090         -120.59314               Falcon 1
#> 2          Florida         28.561857          -80.57737               Falcon 9
#> 3            Texas         25.997264          -97.15608               Falcon 9
#> 4 Marshall Islands          9.047721          167.74313               Falcon 1
#> 5       California         34.632093         -120.61083               Falcon 9
#> 6          Florida         28.608058          -80.60396 Falcon 9, Falcon Heavy
#>   attempted_launches successful_launches
#> 1                  0                   0
#> 2                 52                  50
#> 3                  0                   0
#> 4                  5                   2
#> 5                 15                  15
#> 6                 21                  21
#>                                                                                 wikipedia
#> 1                     https://en.wikipedia.org/wiki/Vandenberg_AFB_Space_Launch_Complex_3
#> 2  https://en.wikipedia.org/wiki/Cape_Canaveral_Air_Force_Station_Space_Launch_Complex_40
#> 3                            https://en.wikipedia.org/wiki/SpaceX_South_Texas_Launch_Site
#> 4                                             https://en.wikipedia.org/wiki/Omelek_Island
#> 5                     https://en.wikipedia.org/wiki/Vandenberg_AFB_Space_Launch_Complex_4
#> 6 https://en.wikipedia.org/wiki/Kennedy_Space_Center_Launch_Complex_39#Launch_Complex_39A
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                    details
#> 1                                                                                                                                                                                                                                                                                                   SpaceX original west coast launch pad for Falcon 1. Performed a static fire but was never used for a launch and abandoned due to scheduling conflicts.
#> 2                                                                                                                                                                                                                                             SpaceX primary Falcon 9 launch pad, where all east coast Falcon 9s launched prior to the AMOS-6 anomaly. Initially used to launch Titan rockets for Lockheed Martin. Back online since CRS-13 on 2017-12-15.
#> 3                                                                                                                                                                                                                               SpaceX new launch site currently under construction to help keep up with the Falcon 9 and Heavy manifests. Expected to be completed in late 2018. Initially will be limited to 12 flights per year, and only GTO launches.
#> 4                                                                                                                                                                                                                                                                                                      SpaceX original launch site, where all of the Falcon 1 launches occured. Abandoned as SpaceX decided against upgrading the pad to support Falcon 9.
#> 5                                                                                                                                                                                                                                                                                     SpaceX primary west coast launch pad for polar orbits and sun synchronous orbits, primarily used for Iridium. Also intended to be capable of launching Falcon Heavy.
#> 6 NASA historic launch pad that launched most of the Saturn V and Space Shuttle missions. Initially for Falcon Heavy launches, it is now launching all of SpaceX east coast missions due to the damage from the AMOS-6 anomaly. After SLC-40 repairs are complete, it will be upgraded to support Falcon Heavy, a process which will take about two months. In the future it will launch commercial crew missions and the Interplanetary Transport System.
#>           site_id                                           site_name_long
#> 1     vafb_slc_3w        Vandenberg Air Force Base Space Launch Complex 3W
#> 2    ccafs_slc_40 Cape Canaveral Air Force Station Space Launch Complex 40
#> 3            stls                           SpaceX South Texas Launch Site
#> 4 kwajalein_atoll                            Kwajalein Atoll Omelek Island
#> 5     vafb_slc_4e        Vandenberg Air Force Base Space Launch Complex 4E
#> 6      ksc_lc_39a         Kennedy Space Center Historic Launch Complex 39A

## get details about launchpad "ksc_lc_39a"
get_SpaceX_launchpads(site_id = "ksc_lc_39a")

#>   id            name             status             location.name
#> 1  5     VAFB SLC 3W            retired Vandenberg Air Force Base
#> 2  2    CCAFS SLC 40             active            Cape Canaveral
#> 3  8            STLS under construction        Boca Chica Village
#> 4  1 Kwajalein Atoll            retired             Omelek Island
#> 5  6     VAFB SLC 4E             active Vandenberg Air Force Base
#> 6  4      KSC LC 39A             active            Cape Canaveral
#>    location.region location.latitude location.longitude      vehicles_launched
#> 1       California         34.644090         -120.59314               Falcon 1
#> 2          Florida         28.561857          -80.57737               Falcon 9
#> 3            Texas         25.997264          -97.15608               Falcon 9
#> 4 Marshall Islands          9.047721          167.74313               Falcon 1
#> 5       California         34.632093         -120.61083               Falcon 9
#> 6          Florida         28.608058          -80.60396 Falcon 9, Falcon Heavy
#>   attempted_launches successful_launches
#> 1                  0                   0
#> 2                 52                  50
#> 3                  0                   0
#> 4                  5                   2
#> 5                 15                  15
#> 6                 21                  21
#>                                                                                 wikipedia
#> 1                     https://en.wikipedia.org/wiki/Vandenberg_AFB_Space_Launch_Complex_3
#> 2  https://en.wikipedia.org/wiki/Cape_Canaveral_Air_Force_Station_Space_Launch_Complex_40
#> 3                            https://en.wikipedia.org/wiki/SpaceX_South_Texas_Launch_Site
#> 4                                             https://en.wikipedia.org/wiki/Omelek_Island
#> 5                     https://en.wikipedia.org/wiki/Vandenberg_AFB_Space_Launch_Complex_4
#> 6 https://en.wikipedia.org/wiki/Kennedy_Space_Center_Launch_Complex_39#Launch_Complex_39A
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                    details
#> 1                                                                                                                                                                                                                                                                                                   SpaceX original west coast launch pad for Falcon 1. Performed a static fire but was never used for a launch and abandoned due to scheduling conflicts.
#> 2                                                                                                                                                                                                                                             SpaceX primary Falcon 9 launch pad, where all east coast Falcon 9s launched prior to the AMOS-6 anomaly. Initially used to launch Titan rockets for Lockheed Martin. Back online since CRS-13 on 2017-12-15.
#> 3                                                                                                                                                                                                                               SpaceX new launch site currently under construction to help keep up with the Falcon 9 and Heavy manifests. Expected to be completed in late 2018. Initially will be limited to 12 flights per year, and only GTO launches.
#> 4                                                                                                                                                                                                                                                                                                      SpaceX original launch site, where all of the Falcon 1 launches occured. Abandoned as SpaceX decided against upgrading the pad to support Falcon 9.
#> 5                                                                                                                                                                                                                                                                                     SpaceX primary west coast launch pad for polar orbits and sun synchronous orbits, primarily used for Iridium. Also intended to be capable of launching Falcon Heavy.
#> 6 NASA historic launch pad that launched most of the Saturn V and Space Shuttle missions. Initially for Falcon Heavy launches, it is now launching all of SpaceX east coast missions due to the damage from the AMOS-6 anomaly. After SLC-40 repairs are complete, it will be upgraded to support Falcon Heavy, a process which will take about two months. In the future it will launch commercial crew missions and the Interplanetary Transport System.
#>           site_id                                           site_name_long
#> 1     vafb_slc_3w        Vandenberg Air Force Base Space Launch Complex 3W
#> 2    ccafs_slc_40 Cape Canaveral Air Force Station Space Launch Complex 40
#> 3            stls                           SpaceX South Texas Launch Site
#> 4 kwajalein_atoll                            Kwajalein Atoll Omelek Island
#> 5     vafb_slc_4e        Vandenberg Air Force Base Space Launch Complex 4E
#> 6      ksc_lc_39a         Kennedy Space Center Historic Launch Complex 39A