GET detailed info about SpaceX parts

GET detailed info about SpaceX parts

get_SpaceX_parts(type, name = NULL, ...)

Arguments

type
character: Type of parts. Choose between "caps" for capsule details or "cores" for core details.

name
character: Name of capsule or core.

...
: Query string to API. See more in details.

Value

data.frame with parsed data from SpaceX API.

Details

Query Strings for type = "caps"

Query Strings	Description

capsule_serial	Filter by capsule serial #

capsule_id	Filter by capsule id

status	Filter by capsule status

original_launch	Filter by original launch date

missions	Filter by flight missions

landings	Filter by # of landings

type	Filter by type of dragon capusle

Query Strings for type = "cores"

core_serial	Filter by core serial #

status	Filter by flight status

original_launch	Filter by original launch date

missions	Filter by flight missions

rtls_attempt	Filter by at least 1 rtls attempt

rtls_landings	Filter by total number of landings

asds_attempt	Filter by at least 1 asds attempt

asds_landings	Filter by total number of landings

water_landing	Filter by at least 1 water landing

Examples

## get details about all cores
get_SpaceX_parts(type = "cores")

#>    core_serial block   status          original_launch original_launch_unix
#> 1     Merlin1A    NA     lost 2006-03-24T22:30:00.000Z           1143239400
#> 2     Merlin2A    NA     lost 2007-03-21T01:10:00.000Z           1174439400
#> 3     Merlin1C    NA     lost 2008-08-02T03:34:00.000Z           1217648040
#> 4     Merlin2C    NA     lost 2008-09-28T23:15:00.000Z           1222643700
#> 5     Merlin3C    NA     lost 2009-07-13T03:35:00.000Z           1247456100
#> 6        B0003     1     lost 2010-06-04T18:45:00.000Z           1275677100
#> 7        B0004     1     lost 2010-12-08T15:43:00.000Z           1291822980
#> 8        B0005     1     lost 2012-05-22T07:44:00.000Z           1335944640
#> 9        B0006     1     lost 2012-10-08T00:35:00.000Z           1349656500
#> 10       B0007     1     lost 2013-03-01T19:10:00.000Z           1362165000
#> 11       B1003     1     lost 2013-09-29T16:00:00.000Z           1380470400
#> 12       B1004     1     lost 2013-12-03T22:41:00.000Z           1386110460
#> 13       B1005     1     lost 2014-01-06T18:06:00.000Z           1389031560
#> 14       B1006     1     lost 2014-04-18T19:25:00.000Z           1397849100
#> 15       B1007     1     lost 2014-07-14T15:15:00.000Z           1405350900
#> 16       B1008     1     lost 2014-08-05T08:00:00.000Z           1407225600
#> 17       B1011     1     lost 2014-09-07T05:00:00.000Z           1410066000
#> 18       B1010     1     lost 2014-09-21T05:52:00.000Z           1411278720
#> 19       B1012     1     lost 2015-01-10T09:47:00.000Z           1420883220
#> 20       B1013     1     lost 2015-02-11T23:03:00.000Z           1423695780
#> 21       B1014     1     lost 2015-03-02T03:50:00.000Z           1425268200
#> 22       B1015     1     lost 2015-04-14T20:10:00.000Z           1429042200
#> 23       B1016     1     lost 2015-04-27T23:03:00.000Z           1430175780
#> 24       B1018     1     lost 2015-06-28T14:21:00.000Z           1435501260
#> 25       B1019     1 inactive 2015-12-22T01:29:00.000Z           1450747740
#> 26       B1017     1     lost 2016-01-17T15:42:00.000Z           1453045320
#> 27       B1020     1     lost 2016-03-04T23:35:00.000Z           1457134500
#> 28       B1021     2 inactive 2016-04-08T20:43:00.000Z           1460148180
#> 29       B1022     2 inactive 2016-05-06T05:21:00.000Z           1462512060
#> 30       B1023     2 inactive 2016-05-27T21:39:00.000Z           1464385140
#> 31       B1024     2     lost 2016-06-15T14:29:00.000Z           1466000940
#> 32       B1025     2 inactive 2016-07-18T04:45:00.000Z           1468817100
#> 33       B1026     2 inactive 2016-08-14T05:26:00.000Z           1471152360
#> 34       B1028     3     lost 2016-09-01T13:07:00.000Z           1472735220
#> 35       B1029     3 inactive 2017-01-14T17:54:00.000Z           1484416440
#> 36       B1031     3 inactive 2017-02-19T14:39:00.000Z           1487515140
#> 37       B1030     3     lost 2017-03-16T06:00:00.000Z           1489644000
#> 38       B1032     3     lost 2017-05-01T11:15:00.000Z           1493637300
#> 39       B1034     3     lost 2017-05-15T23:21:00.000Z           1494890460
#> 40       B1035     3 inactive 2017-06-03T21:07:00.000Z           1496524020
#> 41       B1036     3     lost 2017-06-25T20:25:00.000Z           1498422300
#> 42       B1037     3     lost 2017-07-05T23:35:00.000Z           1499297700
#> 43       B1039     4     lost 2017-08-14T16:31:00.000Z           1502728260
#> 44       B1038     3     lost 2017-08-24T18:50:00.000Z           1503600600
#> 45       B1040     4     lost 2017-09-07T13:50:00.000Z           1504792200
#> 46       B1041     4     lost 2017-10-09T12:37:00.000Z           1507552620
#> 47       B1042     4 inactive 2017-10-30T19:34:00.000Z           1509392040
#> 48       B1043     4     lost 2018-01-08T01:00:00.000Z           1515373200
#> 49       B1033     3     lost 2018-02-06T20:45:00.000Z           1517949900
#> 50       B1044     4     lost 2018-03-06T05:33:00.000Z           1520314380
#> 51       B1045     4     lost 2018-04-18T22:51:00.000Z           1524091860
#> 52       B1046     5     lost 2018-05-11T20:14:00.000Z           1526069640
#> 53       B1047     5     lost 2018-07-22T05:50:00.000Z           1532238600
#> 54       B1048     5     lost 2018-07-25T11:39:26.000Z           1532518766
#> 55       B1049     5   active 2018-09-10T04:45:00.000Z           1536554700
#> 56       B1050     5 inactive 2018-12-05T18:16:00.000Z           1544033760
#> 57       B1054     5     lost 2018-12-23T13:51:00.000Z           1545573060
#> 58       B1051     5  unknown 2019-03-02T07:45:00.000Z           1551512700
#> 59       B1052     5  unknown 2019-04-11T22:35:00.000Z           1555022100
#> 60       B1053     5  unknown 2019-04-11T22:35:00.000Z           1555022100
#> 61       B1055     5     lost 2019-04-11T22:35:00.000Z           1555022100
#> 62       B1056     5     lost 2019-05-04T06:48:00.000Z           1556952480
#> 63       B1057     5     lost 2019-06-25T03:30:00.000Z           1561433400
#> 64       B1059     5  unknown 2019-12-05T17:29:23.000Z           1575566963
#> 65       B1058     5  unknown 2020-05-30T19:22:00.000Z           1590866520
#>                                                                                                                  missions
#> 1                                                                                                            FalconSat, 1
#> 2                                                                                                              DemoSat, 2
#> 3                                                                                                          Trailblazer, 3
#> 4                                                                                                               RatSat, 4
#> 5                                                                                                             RazakSat, 5
#> 6                                                                                                 Falcon 9 Test Flight, 6
#> 7                                                                                                               COTS 1, 7
#> 8                                                                                                               COTS 2, 8
#> 9                                                                                                                CRS-1, 9
#> 10                                                                                                              CRS-2, 10
#> 11                                                                                                           CASSIOPE, 11
#> 12                                                                                                              SES-8, 12
#> 13                                                                                                          Thaicom 6, 13
#> 14                                                                                                              CRS-3, 14
#> 15                                                                                                     OG-2 Mission 1, 15
#> 16                                                                                                          AsiaSat 8, 16
#> 17                                                                                                          AsiaSat 6, 17
#> 18                                                                                                              CRS-4, 18
#> 19                                                                                                              CRS-5, 19
#> 20                                                                                                             DSCOVR, 20
#> 21                                                                                           ABS-3A / Eutelsat 115W B, 21
#> 22                                                                                                              CRS-6, 22
#> 23                                                                                       TürkmenÄlem 52°E / MonacoSAT, 23
#> 24                                                                                                              CRS-7, 24
#> 25                                                                                                     OG-2 Mission 2, 25
#> 26                                                                                                            Jason 3, 26
#> 27                                                                                                              SES-9, 27
#> 28                                                                                                  CRS-8, SES-10, 28, 38
#> 29                                                                                                           JCSAT-2B, 29
#> 30                                                                            Thaicom 8, Falcon Heavy Test Flight, 30, 55
#> 31                                                                                           ABS-2A / Eutelsat 117W B, 31
#> 32                                                                                CRS-9, Falcon Heavy Test Flight, 32, 55
#> 33                                                                                                           JCSAT-16, 33
#> 34                                                                                                             Amos-6, 34
#> 35                                                                          Iridium NEXT Mission 1, BulgariaSat-1, 35, 42
#> 36                                                                                  CRS-10, SES-11 / Echostar 105, 36, 49
#> 37                                                                                                        EchoStar 23, 37
#> 38                                                                                     NROL-76, SES-16 / GovSat-1, 39, 54
#> 39                                                                                                      Inmarsat-5 F4, 40
#> 40                                                                                                 CRS-11, CRS-13, 41, 51
#> 41                                                                 Iridium NEXT Mission 2, Iridium NEXT Mission 4, 43, 52
#> 42                                                                                                       Intelsat 35e, 44
#> 43                                                                                                 CRS-12, CRS-14, 45, 59
#> 44                                                                                FormoSat-5, Paz / Starlink Demo, 46, 56
#> 45                                                                                     Boeing X-37B OTV-5, SES-12, 47, 63
#> 46                                                                 Iridium NEXT Mission 3, Iridium NEXT Mission 5, 48, 58
#> 47                                                                                                        KoreaSat 5A, 50
#> 48                                                                                   ZUMA, Iridium NEXT Mission 6, 53, 62
#> 49                                                                                           Falcon Heavy Test Flight, 55
#> 50                                                                                                     Hispasat 30W-6, 57
#> 51                                                                                                   TESS, CRS-15, 60, 64
#> 52                                    Bangabandhu-1, Merah Putih, SSO-A, Crew Dragon In Flight Abort Test, 61, 67, 71, 88
#> 53                                                                            Telstar 19V, Es’hail 2, Amos-17, 65, 70, 83
#> 54 Iridium NEXT Mission 7, SAOCOM 1A, Nusantara Satu (PSN-6) / S5 / Beresheet, Starlink 1, Starlink 5, 66, 69, 75, 84, 92
#> 55                         Telstar 18V, Iridium NEXT Mission 8, Starlink v0.9, Starlink 2, Starlink 7, 68, 74, 79, 87, 95
#> 56                                                                                                             CRS-16, 72
#> 57                                                                                                       GPS III SV01, 73
#> 58                                  CCtCap Demo Mission 1, RADARSAT Constellation, Starlink 3, Starlink 6, 76, 80, 89, 93
#> 59                                                                                              ArabSat 6A, STP-2, 77, 81
#> 60                                                                                              ArabSat 6A, STP-2, 77, 81
#> 61                                                                                                         ArabSat 6A, 77
#> 62                                                       CRS-17, CRS-18, JCSat 18 / Kacific 1, Starlink 4, 78, 82, 86, 90
#> 63                                                                                                              STP-2, 81
#> 64                                                                                                 CRS-19, CRS-20, 85, 91
#> 65                                                                                              CCtCap Demo Mission 2, 94
#>    reuse_count rtls_attempts rtls_landings asds_attempts asds_landings
#> 1            0             0             0             0             0
#> 2            0             0             0             0             0
#> 3            0             0             0             0             0
#> 4            0             0             0             0             0
#> 5            0             0             0             0             0
#> 6            0             0             0             0             0
#> 7            0             0             0             0             0
#> 8            0             0             0             0             0
#> 9            0             0             0             0             0
#> 10           0             0             0             0             0
#> 11           0             0             0             0             0
#> 12           0             0             0             0             0
#> 13           0             0             0             0             0
#> 14           0             0             0             0             0
#> 15           0             0             0             0             0
#> 16           0             0             0             0             0
#> 17           0             0             0             0             0
#> 18           0             0             0             0             0
#> 19           0             0             0             1             0
#> 20           0             0             0             0             0
#> 21           0             0             0             0             0
#> 22           0             0             0             1             0
#> 23           0             0             0             0             0
#> 24           0             0             0             0             0
#> 25           0             1             1             0             0
#> 26           0             0             0             1             0
#> 27           0             0             0             1             0
#> 28           1             0             0             2             2
#> 29           0             0             0             1             1
#> 30           1             1             1             1             1
#> 31           0             0             0             1             0
#> 32           1             2             2             0             0
#> 33           0             0             0             1             1
#> 34           0             0             0             0             0
#> 35           1             0             0             2             2
#> 36           1             1             1             1             1
#> 37           0             0             0             0             0
#> 38           1             1             1             0             0
#> 39           0             0             0             0             0
#> 40           1             2             2             0             0
#> 41           1             0             0             1             1
#> 42           0             0             0             0             0
#> 43           1             1             1             0             0
#> 44           1             0             0             1             1
#> 45           1             1             1             0             0
#> 46           1             0             0             1             1
#> 47           0             0             0             1             1
#> 48           1             1             1             0             0
#> 49           0             0             0             1             0
#> 50           0             0             0             0             0
#> 51           1             0             0             1             1
#> 52           3             0             0             3             3
#> 53           2             0             0             2             2
#> 54           4             1             1             4             3
#> 55           4             0             0             5             5
#> 56           0             1             0             0             0
#> 57           0             0             0             0             0
#> 58           3             1             1             3             3
#> 59           1             2             2             0             0
#> 60           1             2             2             0             0
#> 61           0             0             0             1             1
#> 62           3             1             1             3             2
#> 63           0             0             0             1             0
#> 64           1             1             1             1             1
#> 65           0             0             0             1             1
#>    water_landing
#> 1          FALSE
#> 2          FALSE
#> 3          FALSE
#> 4          FALSE
#> 5          FALSE
#> 6          FALSE
#> 7          FALSE
#> 8          FALSE
#> 9          FALSE
#> 10         FALSE
#> 11          TRUE
#> 12         FALSE
#> 13         FALSE
#> 14          TRUE
#> 15          TRUE
#> 16         FALSE
#> 17         FALSE
#> 18          TRUE
#> 19         FALSE
#> 20          TRUE
#> 21         FALSE
#> 22         FALSE
#> 23         FALSE
#> 24         FALSE
#> 25         FALSE
#> 26         FALSE
#> 27         FALSE
#> 28         FALSE
#> 29         FALSE
#> 30         FALSE
#> 31         FALSE
#> 32         FALSE
#> 33         FALSE
#> 34         FALSE
#> 35         FALSE
#> 36         FALSE
#> 37         FALSE
#> 38          TRUE
#> 39         FALSE
#> 40         FALSE
#> 41          TRUE
#> 42         FALSE
#> 43         FALSE
#> 44         FALSE
#> 45         FALSE
#> 46         FALSE
#> 47         FALSE
#> 48         FALSE
#> 49         FALSE
#> 50         FALSE
#> 51         FALSE
#> 52         FALSE
#> 53         FALSE
#> 54         FALSE
#> 55         FALSE
#> 56          TRUE
#> 57         FALSE
#> 58         FALSE
#> 59         FALSE
#> 60         FALSE
#> 61         FALSE
#> 62         FALSE
#> 63         FALSE
#> 64         FALSE
#> 65         FALSE
#>                                                                                                                                                                                       details
#> 1                                                                                                                                  Engine failure at T+33 seconds resulted in loss of vehicle
#> 2  Successful first-stage burn and transition to second stage, maximal altitude 289 km. Harmonic oscillation at T+5 minutes Premature engine shutdown at T+7 min 30 s. Failed to reach orbit.
#> 3                                                                                                                       Residual stage-1 thrust led to collision between stage 1 and stage 2.
#> 4                                                                      Initially scheduled for 23–25 Sep, carried dummy payload – mass simulator, 165 kg (originally intended to be RazakSAT.
#> 5                                                                                                                                                                                        <NA>
#> 6                                                                                                                                                                              Stage Expended
#> 7                                                                                                                                                                      First flight of Dragon
#> 8                                                                                                                                                                                        <NA>
#> 9                                                                                                                                Suffered engine out at T+1:19 but primary mission successful
#> 10                                                                                                                                                                                       <NA>
#> 11                                                                                                                 First flight of Falcon 9 v1.1 upgrade, first Spacex flight from Vandenberg
#> 12                                                                                                                                                                                       <NA>
#> 13                                                                                                                                                                                       <NA>
#> 14                                                                                                                                                     Broke up after sucessful water landing
#> 15                                                                                                                                                     Broke up after sucessful water landing
#> 16                                                                                                                                                                                       <NA>
#> 17                                                                                                                                                                                       <NA>
#> 18                                                                                                                                                     Broke up after sucessful water landing
#> 19                                                                                                                      Destroyed on impact with droneship, grid fin hydraulic fluid depleted
#> 20                                                                                                                                                     Broke up after sucessful water landing
#> 21                                                                                                                                                                                       <NA>
#> 22                                                                                                      Destroyed on impact with droneship due to excess lateral velocity - throttle stiction
#> 23                                                                                                                                                                                       <NA>
#> 24                                                                                                   RUD at T + 139s - second stage overpressure event due to defective COPV strut heim joint
#> 25                                                                                                                                 Retired; permanently displayed at Hawthorne headquarters. 
#> 26                                                                                                                              Destroyed after drone ship landing due to landing leg failure
#> 27                                                                                                    Destroyed on Impact with DroneShip (OCISLY) - Excess Velocity, Possible Fuel Exhaustion
#> 28                                                                                                                                                Core will be given as a gift to the Cape.  
#> 29                                              Removed from active fleet for Δ Qualification Testing. Retired after eight to ten test cycles.  Currently partially disassembled at McGregor.
#> 30                                                                                                                                                                     *At LZ1 as 2018/02/06 
#> 31                                                                                                                                                          RUD on deck of droneship (OCISLY)
#> 32                                                                                                                                                                     *At LZ2 as 2018/02/06 
#> 33                                                                                                                                                                    Permanently scrapped.  
#> 34                                                                                                                                                                                        ---
#> 35                                                                                                                                                     Will be given to KSC Visitors Center. 
#> 36                                                                                                                     Stored outside at McGregor, partially disassembled (missing engines). 
#> 37                                                                                                                                                                             Stage Expended
#> 38                                                                                      Survived water landing, SpaceX couldn't tug it back to port, so they hired a company to destroy it.  
#> 39                                                                                                                                                                             Stage Expended
#> 40                                                                                                                               To be moved to Space Center Houston for display this summer.
#> 41                                                                                                                                                    Broke Up After Successful Water Landing
#> 42                                                                                                                                                                             Stage Expended
#> 43                                                                                                                                                                             Stage Expended
#> 44                                                                                                                                                                             Stage Expended
#> 45                                                                                                                                                                             Stage Expended
#> 46                                                                                                                                                                             Stage Expended
#> 47                                                                                                                                                                                  Retired. 
#> 48                                                                                                                                                                             Stage Expended
#> 49                                                                                              Two of the outer engines didn't re-ignite for the landing burn, ran out of TEA/TEB igniters. 
#> 50                                                                                                                                                                             Stage Expended
#> 51                                                                                                                                                                             Stage Expended
#> 52                                                                                                                                                                             Stage Expended
#> 53                                                                                                                                                                             Stage Expended
#> 54                                                                                                                   Landing burn not performed after premature engine shutdown during ascent
#> 55                                                                                                                                                                 At SLC-40 awaiting launch.
#> 56                                                                                                                   Towed to Port of Canaveral; cannibalized for parts to build Starhopper. 
#> 57                                                                                                                                                                             Stage Expended
#> 58                                                                                                                                                      Landed on OCISLY as of Apr 22, 2020. 
#> 59                                                                                                                                                        Landed at LZ-1 as of Jun 25, 2019. 
#> 60                                                                                                                                                        Landed at LZ-2 as of Jun 25, 2019. 
#> 61                                                                                                                 Toppled by rough seas and broken in two after safely landing aboard OCISLY
#> 62                                                                                             Missed the droneship and made successful water landing; apparently scuttled at sea afterward. 
#> 63                                                                                                                                     Exploded upon ocean impact a safe distance from OCISLY
#> 64                                                                                                                                                         Landed at LZ-1 as of Mar 7, 2020. 
#> 65                                                                                                                                                       Landed on OCISLY as of May 30, 2020 

## get details about all capsules
get_SpaceX_parts(type = "caps")

#>    capsule_serial capsule_id    status          original_launch
#> 1            C202    dragon2    active                     <NA>
#> 2            C203    dragon2    active                     <NA>
#> 3            C204    dragon2    active                     <NA>
#> 4            C101    dragon1   retired 2010-12-08T15:43:00.000Z
#> 5            C102    dragon1   retired 2012-05-22T07:44:00.000Z
#> 6            C103    dragon1   unknown 2012-10-08T00:35:00.000Z
#> 7            C104    dragon1   unknown 2013-03-01T19:10:00.000Z
#> 8            C105    dragon1   unknown 2014-04-18T19:25:00.000Z
#> 9            C106    dragon1    active 2014-09-21T05:52:00.000Z
#> 10           C107    dragon1   unknown 2015-01-10T09:47:00.000Z
#> 11           C108    dragon1    active 2015-04-14T20:10:00.000Z
#> 12           C109    dragon1 destroyed 2015-06-28T14:21:00.000Z
#> 13           C110    dragon1    active 2016-04-08T20:43:00.000Z
#> 14           C111    dragon1    active 2016-07-18T04:45:00.000Z
#> 15           C112    dragon1    active 2017-02-19T14:39:00.000Z
#> 16           C113    dragon1    active 2017-08-14T16:31:00.000Z
#> 17           C201    dragon2    active 2019-03-02T07:45:00.000Z
#> 18           C205    dragon2    active 2020-01-19T14:00:00.000Z
#> 19           C206    dragon2    active 2020-05-30T19:22:00.000Z
#>    original_launch_unix                             missions landings
#> 1                    NA                                 NULL        0
#> 2                    NA                                 NULL        0
#> 3                    NA                                 NULL        0
#> 4            1291822980                            COTS 1, 7        0
#> 5            1335944640                            COTS 2, 8        1
#> 6            1349656500                             CRS-1, 9        1
#> 7            1362165000                            CRS-2, 10        1
#> 8            1397849100                            CRS-3, 14        1
#> 9            1411278720    CRS-4, CRS-11, CRS-19, 18, 41, 85        3
#> 10           1420883220                            CRS-5, 19        1
#> 11           1429042200    CRS-6, CRS-13, CRS-18, 22, 51, 82        3
#> 12           1435501260                            CRS-7, 24        0
#> 13           1460148180                CRS-8, CRS-14, 28, 59        2
#> 14           1468817100                CRS-9, CRS-15, 32, 64        2
#> 15           1487515140   CRS-10, CRS-16, CRS-20, 36, 72, 91        3
#> 16           1502728260               CRS-12, CRS-17, 45, 78        2
#> 17           1551512700            CCtCap Demo Mission 1, 76        1
#> 18           1579442400 Crew Dragon In Flight Abort Test, 88        1
#> 19           1590866520            CCtCap Demo Mission 2, 94        1
#>          type
#> 1  Dragon 2.0
#> 2  Dragon 2.0
#> 3  Dragon 2.0
#> 4  Dragon 1.0
#> 5  Dragon 1.0
#> 6  Dragon 1.0
#> 7  Dragon 1.0
#> 8  Dragon 1.1
#> 9  Dragon 1.1
#> 10 Dragon 1.1
#> 11 Dragon 1.1
#> 12 Dragon 1.1
#> 13 Dragon 1.1
#> 14 Dragon 1.1
#> 15 Dragon 1.1
#> 16 Dragon 1.1
#> 17 Dragon 2.0
#> 18 Dragon 2.0
#> 19 Dragon 2.0
#>                                                                                                  details
#> 1                     Capsule used to qualify Dragon 2's environmental control and life support systems.
#> 2                                                  Rumored to be used for Inflight Abort Test after DM-1
#> 3                                                              Currently in construction for use in DM-2
#> 4                                                                   Reentered after three weeks in orbit
#> 5                                                                                First Dragon spacecraft
#> 6                                                 First of twenty missions flown under the CRS1 contract
#> 7                                                                                                   <NA>
#> 8                                                                              First Dragon v1.1 capsule
#> 9                                                                      First Dragon capsule to be reused
#> 10                                                                                                  <NA>
#> 11                                                                    Second Dragon capsule to be reused
#> 12                                                           Destroyed on impact after F9 launch failure
#> 13                                                                                                  <NA>
#> 14                                                                                                  <NA>
#> 15                                                                                                  <NA>
#> 16                                                                  The last newly manufactured Dragon 1
#> 17                                                 Destroyed in a test anomaly at LZ-1 on April 20, 2019
#> 18                        In construction for use in first mission in contract under the CCtCap contract
#> 19 In Hawthorne, CA as of October 11, 2019. Seen in background during Musk/Bridenstine press conference.
#>    reuse_count
#> 1            0
#> 2            0
#> 3            0
#> 4            0
#> 5            0
#> 6            0
#> 7            0
#> 8            0
#> 9            2
#> 10           0
#> 11           2
#> 12           0
#> 13           1
#> 14           1
#> 15           2
#> 16           1
#> 17           0
#> 18           0
#> 19           0

## get details about all active capsules
get_SpaceX_parts(type = "caps", status = "active")

#>    capsule_serial capsule_id status          original_launch
#> 1            C202    dragon2 active                     <NA>
#> 2            C203    dragon2 active                     <NA>
#> 3            C204    dragon2 active                     <NA>
#> 4            C106    dragon1 active 2014-09-21T05:52:00.000Z
#> 5            C108    dragon1 active 2015-04-14T20:10:00.000Z
#> 6            C110    dragon1 active 2016-04-08T20:43:00.000Z
#> 7            C111    dragon1 active 2016-07-18T04:45:00.000Z
#> 8            C112    dragon1 active 2017-02-19T14:39:00.000Z
#> 9            C113    dragon1 active 2017-08-14T16:31:00.000Z
#> 10           C201    dragon2 active 2019-03-02T07:45:00.000Z
#> 11           C205    dragon2 active 2020-01-19T14:00:00.000Z
#> 12           C206    dragon2 active 2020-05-30T19:22:00.000Z
#>    original_launch_unix                             missions landings
#> 1                    NA                                 NULL        0
#> 2                    NA                                 NULL        0
#> 3                    NA                                 NULL        0
#> 4            1411278720    CRS-4, CRS-11, CRS-19, 18, 41, 85        3
#> 5            1429042200    CRS-6, CRS-13, CRS-18, 22, 51, 82        3
#> 6            1460148180                CRS-8, CRS-14, 28, 59        2
#> 7            1468817100                CRS-9, CRS-15, 32, 64        2
#> 8            1487515140   CRS-10, CRS-16, CRS-20, 36, 72, 91        3
#> 9            1502728260               CRS-12, CRS-17, 45, 78        2
#> 10           1551512700            CCtCap Demo Mission 1, 76        1
#> 11           1579442400 Crew Dragon In Flight Abort Test, 88        1
#> 12           1590866520            CCtCap Demo Mission 2, 94        1
#>          type
#> 1  Dragon 2.0
#> 2  Dragon 2.0
#> 3  Dragon 2.0
#> 4  Dragon 1.1
#> 5  Dragon 1.1
#> 6  Dragon 1.1
#> 7  Dragon 1.1
#> 8  Dragon 1.1
#> 9  Dragon 1.1
#> 10 Dragon 2.0
#> 11 Dragon 2.0
#> 12 Dragon 2.0
#>                                                                                                  details
#> 1                     Capsule used to qualify Dragon 2's environmental control and life support systems.
#> 2                                                  Rumored to be used for Inflight Abort Test after DM-1
#> 3                                                              Currently in construction for use in DM-2
#> 4                                                                      First Dragon capsule to be reused
#> 5                                                                     Second Dragon capsule to be reused
#> 6                                                                                                   <NA>
#> 7                                                                                                   <NA>
#> 8                                                                                                   <NA>
#> 9                                                                   The last newly manufactured Dragon 1
#> 10                                                 Destroyed in a test anomaly at LZ-1 on April 20, 2019
#> 11                        In construction for use in first mission in contract under the CCtCap contract
#> 12 In Hawthorne, CA as of October 11, 2019. Seen in background during Musk/Bridenstine press conference.
#>    reuse_count
#> 1            0
#> 2            0
#> 3            0
#> 4            2
#> 5            2
#> 6            1
#> 7            1
#> 8            2
#> 9            1
#> 10           0
#> 11           0
#> 12           0

## get details about core B1041
get_SpaceX_parts(type = "cores", name = "B1041")

#> $core_serial
#> [1] "B1041"
#> 
#> $block
#> [1] 4
#> 
#> $status
#> [1] "lost"
#> 
#> $original_launch
#> [1] "2017-10-09T12:37:00.000Z"
#> 
#> $original_launch_unix
#> [1] 1507552620
#> 
#> $missions
#>                     name flight
#> 1 Iridium NEXT Mission 3     48
#> 2 Iridium NEXT Mission 5     58
#> 
#> $reuse_count
#> [1] 1
#> 
#> $rtls_attempts
#> [1] 0
#> 
#> $rtls_landings
#> [1] 0
#> 
#> $asds_attempts
#> [1] 1
#> 
#> $asds_landings
#> [1] 1
#> 
#> $water_landing
#> [1] FALSE
#> 
#> $details
#> [1] "Stage Expended"
#>