GET detailed info about SpaceX parts

GET detailed info about SpaceX parts

get_SpaceX_parts(type, name = NULL, ...)

Arguments

type
character: Type of parts. Choose between "caps" for capsule details or "cores" for core details.

name
character: Name of capsule or core.

...
: Query string to API. See more in details.

Value

data.frame with parsed data from SpaceX API.

Details

Query Strings for type = "caps"

Query Strings 	 Description

capsule_serial 	 Filter by capsule serial #

capsule_id 	 Filter by capsule id

status 	 Filter by capsule status

original_launch 	 Filter by original launch date

missions 	 Filter by flight missions

landings 	 Filter by # of landings

type 	 Filter by type of dragon capusle

Query Strings for type = "cores"

core_serial 	 Filter by core serial #

status 	 Filter by flight status

original_launch 	 Filter by original launch date

missions 	 Filter by flight missions

rtls_attempt 	 Filter by at least 1 rtls attempt

rtls_landings 	 Filter by total number of landings

asds_attempt 	 Filter by at least 1 asds attempt

asds_landings 	 Filter by total number of landings

water_landing 	 Filter by at least 1 water landing

Examples

## get details about all cores
get_SpaceX_parts(type = "cores")

#>    core_serial    status      original_launch
#> 1        B0003  expended           2010-06-04
#> 2        B0004  expended           2010-12-08
#> 3        B0005  expended           2012-05-22
#> 4        B0006  expended           2012-10-08
#> 5        B0007  expended           2013-03-01
#> 6        B1003 destroyed           2013-09-29
#> 7        B1004  expended           2013-12-03
#> 8        B1005  expended           2014-01-06
#> 9        B1006 destroyed           2014-04-18
#> 10       B1007 destroyed           2014-07-14
#> 11       B1008  expended           2014-08-05
#> 12       B1010  expended           2014-09-21
#> 13       B1011  expended           2014-09-07
#> 14       B1012 destroyed           2015-01-10
#> 15       B1013 destroyed           2015-02-11
#> 16       B1014  expended           2015-03-02
#> 17       B1015 destroyed           2015-04-14
#> 18       B1016  expended           2015-04-27
#> 19       B1017 destroyed           2016-01-17
#> 20       B1018 destroyed           2015-06-28
#> 21       B1019   retired           2015-12-22
#> 22       B1020 destroyed           2016-03-04
#> 23       B1021   retired           2016-04-08
#> 24       B1022   retired           2016-05-06
#> 25       B1023    active           2016-05-27
#> 26       B1024 destroyed           2016-06-15
#> 27       B1025    active           2016-07-18
#> 28       B1026    active           2016-08-14
#> 29       B1028 destroyed           2016-09-01
#> 30       B1029   retired           2017-01-14
#> 31       B1030 destroyed           2017-03-16
#> 32       B1031   retired           2017-02-19
#> 33       B1032    active           2017-05-01
#> 34       B1033    active                  TBD
#> 35       B1034  expended           2017-05-15
#> 36       B1035   retired           2017-06-03
#> 37       B1036 destroyed           2017-05-25
#> 38       B1037  expended           2017-07-05
#> 39       B1038    active 2017-08-24T18:50:00Z
#> 40       B1039    active 2017-08-14T16:31:00Z
#> 41       B1040    active 2017-09-07T13:50:00Z
#> 42       B1041    active 2017-10-09T12:37:00Z
#> 43       B1042    active 2017-10-30T19:34:00Z
#> 44       B1043    active                  TBD
#> 45       B1044    active                  TBD
#> 46       B1045    active                  TBD
#>                                        missions rtls_attempt rtls_landings
#> 1                     Dragon Qualification Unit        FALSE             0
#> 2                            COTS Demo Flight 1        FALSE             0
#> 3                            COTS Demo Flight 2        FALSE             0
#> 4                     SpaceX CRS-1, Orbcomm-OG2        FALSE             0
#> 5                                  SpaceX CRS-2        FALSE             0
#> 6                                      CASSIOPE        FALSE             0
#> 7                                         SES-8        FALSE             0
#> 8                                     Thaicom 6        FALSE             0
#> 9                                  SpaceX CRS-3        FALSE             0
#> 10                              Orbcomm-OG2 x 6        FALSE             0
#> 11                                    AsiaSat 8        FALSE             0
#> 12                                 SpaceX CRS-4        FALSE             0
#> 13                                    AsiaSat 6        FALSE             0
#> 14                                 SpaceX CRS-5        FALSE             0
#> 15                                       DSCOVR        FALSE             0
#> 16                  ABS-3A, Eutelsat 115 West B        FALSE             0
#> 17                                 SpaceX CRS-6        FALSE             0
#> 18                 TürkmenÄlem 52°E / MonacoSAT        FALSE             0
#> 19                                      Jason-3        FALSE             0
#> 20                                 SpaceX CRS-7        FALSE             0
#> 21                               Orbcomm OG2 M2         TRUE             1
#> 22                                        SES-9        FALSE             0
#> 23                         SpaceX CRS-8, SES-10        FALSE             0
#> 24                                     JCSAT-2B        FALSE             0
#> 25                                    Thaicom 8        FALSE             0
#> 26                  ABS-2A, Eutelsat 117 West B        FALSE             0
#> 27                                 SpaceX CRS-9         TRUE             1
#> 28                                     JCSAT-16        FALSE             0
#> 29                                       Amos-6        FALSE             0
#> 30                  Iridium NEXT, BulgariaSat-1        FALSE             0
#> 31                                  EchoStar 23        FALSE             0
#> 32          SpaceX CRS-10, Echostar 105, SES-11         TRUE             1
#> 33                            NROL-76, GovSat-1         TRUE             1
#> 34                            Falcon Heavy Demo        FALSE             0
#> 35                                Inmarsat-5 F4        FALSE             0
#> 36                 SpaceX CRS-11, SpaceX CRS-13         TRUE             2
#> 37 Iridium NEXT Flight 2, Iridium NEXT Flight 4        FALSE             0
#> 38                                 Intelsat 35e        FALSE             0
#> 39                                   FormoSat-5        FALSE             0
#> 40                                SpaceX CRS-12         TRUE             1
#> 41                                  X-37B OTV-5         TRUE             1
#> 42                           Iridium NEXT 21-30        FALSE             0
#> 43                                  KoreaSat 5A        FALSE             0
#> 44                                         ZUMA        FALSE             0
#> 45                          Hispasat 30W-6 (1F)        FALSE             0
#> 46                                          TBD        FALSE             0
#>    asds_attempt asds_landings water_landing
#> 1         FALSE             0         FALSE
#> 2         FALSE             0         FALSE
#> 3         FALSE             0         FALSE
#> 4         FALSE             0         FALSE
#> 5         FALSE             0         FALSE
#> 6         FALSE             0          TRUE
#> 7         FALSE             0         FALSE
#> 8         FALSE             0         FALSE
#> 9         FALSE             0          TRUE
#> 10        FALSE             0          TRUE
#> 11        FALSE             0         FALSE
#> 12        FALSE             0          TRUE
#> 13        FALSE             0         FALSE
#> 14         TRUE             0         FALSE
#> 15        FALSE             0          TRUE
#> 16        FALSE             0         FALSE
#> 17         TRUE             0         FALSE
#> 18        FALSE             0         FALSE
#> 19         TRUE             0         FALSE
#> 20        FALSE             0         FALSE
#> 21        FALSE             0         FALSE
#> 22         TRUE             0         FALSE
#> 23         TRUE             2         FALSE
#> 24         TRUE             1         FALSE
#> 25         TRUE             1         FALSE
#> 26         TRUE             0         FALSE
#> 27        FALSE             0         FALSE
#> 28         TRUE             1         FALSE
#> 29        FALSE             0         FALSE
#> 30         TRUE             2         FALSE
#> 31        FALSE             0         FALSE
#> 32         TRUE             1         FALSE
#> 33        FALSE             0         FALSE
#> 34        FALSE             0         FALSE
#> 35        FALSE             0         FALSE
#> 36        FALSE             0         FALSE
#> 37         TRUE             1          TRUE
#> 38        FALSE             0         FALSE
#> 39         TRUE             1         FALSE
#> 40        FALSE             0         FALSE
#> 41        FALSE             0         FALSE
#> 42         TRUE             1         FALSE
#> 43         TRUE             1         FALSE
#> 44        FALSE             0         FALSE
#> 45        FALSE             0         FALSE
#> 46        FALSE             0         FALSE
#>                                                                                                                                               details
#> 1                                                                               Core expended on flight, no recovery effort. First flight of Falcon 9
#> 2                                                                                                                              First flight of Dragon
#> 3                                                                                                                                                <NA>
#> 4                                                                                        Suffered engine out at T+1:19 but primary mission successful
#> 5                                                                                                                                                <NA>
#> 6                                                                          First flight of Falcon 9 v1.1 upgrade, first Spacex flight from Vandenberg
#> 7                                                                                                                                                <NA>
#> 8                                                                                                                                                <NA>
#> 9                                                                                                              Broke up after sucessful water landing
#> 10                                                                                                             Broke up after sucessful water landing
#> 11                                                                                                                                               <NA>
#> 12                                                                                                             Broke up after sucessful water landing
#> 13                                                                                                                                               <NA>
#> 14                                                                              Destroyed on impact with droneship, grid fin hydraulic fluid depleted
#> 15                                                                                                             Broke up after sucessful water landing
#> 16                                                                                                                                               <NA>
#> 17                                                              Destroyed on impact with droneship due to excess lateral velocity - throttle stiction
#> 18                                                                                                                                               <NA>
#> 19                                                                                      Destroyed after drone ship landing due to landing leg failure
#> 20                                                           RUD at T + 139s - second stage overpressure event due to defective COPV strut heim joint
#> 21                                                                                                 First recovered stage, first return to launch site
#> 22                                                               Destroyed on impact with drone ship due to excess velocity, possible fuel exhaustion
#> 23 First drone ship landing, first orbital booster to be reflown, first orbital booster to be landed twice, will be given as a gift to Cape Canaveral
#> 24                                                First recovered booster from a GTO mission, in storage, and due to be retired after ten test cycles
#> 25                                                                                                             Converted to Falcon Heavy side booster
#> 26                                                                                                                          RUD on deck of drone ship
#> 27                                                                                                         Converted to side booster for Falcon Heavy
#> 28                                                                                                              In long term storage outside hanger X
#> 29                                                                                                  Destroyed on launchpad during static fire anomaly
#> 30                                              Second booster to be reflown, Completed sucessful launch of BulgariaSat-1 and ASDS landing on 6/23/17
#> 31                                                                                                                                               <NA>
#> 32                                                                                              Daytime RTLS landing, helped launch CRS-10 to the ISS
#> 33                                                                               Daytime RTLS landing, helped launch first SpaceX mission for the NRO
#> 34                                                                                                                 Sucessful engine test as of 5/9/17
#> 35                                                                                                                                               <NA>
#> 36                                                                                              Daytime RTLS landing, helped launch CRS-11 to the ISS
#> 37                                                                                                                          Destroyed by waterlanding
#> 38                                                                                                Core will be expended on flight, no recovery effort
#> 39                                                                                                                           In storage at Port of LA
#> 40                                                                                                                                 In storage at LZ-1
#> 41                                                                                                                                 In storage at LZ-1
#> 42                                                                                                                                   In refurbishment
#> 43                                                                                                                                   In refurbishment
#> 44                                                                                                   Being transfered to SLC-40 for a January launch.
#> 45                                                                                                                           At McGregor for testing.
#> 46                                                                                                                           At McGregor for testing.

## get details about all capsules
get_SpaceX_parts(type = "caps")

#>    capsule_serial capsule_id    status      original_launch
#> 1            C101    dragon1   retired           2010-12-08
#> 2            C102    dragon1   retired           2012-05-22
#> 3            C103    dragon1   unknown           2012-10-08
#> 4            C104    dragon1   unknown           2013-03-01
#> 5            C105    dragon1   unknown           2014-04-18
#> 6            C106    dragon1    active           2014-09-21
#> 7            C107    dragon1   unknown           2015-01-10
#> 8            C108    dragon1    active           2015-04-14
#> 9            C109    dragon1 destroyed           2015-06-28
#> 10           C110    dragon1   unknown           2016-04-08
#> 11           C111    dragon1    active           2016-07-18
#> 12           C112    dragon1    active           2017-02-19
#> 13           C113    dragon1    active 2017-08-14T16:31:00Z
#> 14           C201    dragon2    active                  TBD
#> 15           C202    dragon2    active                  TBD
#> 16           C203    dragon2    active                  TBD
#> 17           C204    dragon2    active                  TBD
#> 18           C205    dragon2    active                  TBD
#>                       missions landings       type
#> 1           COTS Demo Flight 1        0 Dragon 1.0
#> 2           COTS Demo Flight 2        1 Dragon 1.0
#> 3                 SpaceX CRS-1        1 Dragon 1.0
#> 4                 SpaceX CRS-2        1 Dragon 1.0
#> 5                 SpaceX CRS-3        1 Dragon 1.1
#> 6  SpaceX CRS-4, SpaceX CRS-11        2 Dragon 1.1
#> 7                 SpaceX CRS-5        1 Dragon 1.1
#> 8  SpaceX CRS-6, SpaceX CRS-13        2 Dragon 1.1
#> 9                 SpaceX CRS-7        0 Dragon 1.1
#> 10                SpaceX CRS-8        1 Dragon 1.1
#> 11                SpaceX CRS-9        1 Dragon 1.1
#> 12               SpaceX CRS-10        1 Dragon 1.1
#> 13               SpaceX CRS-12        1 Dragon 1.1
#> 14                         TBD        0 Dragon 2.0
#> 15                         TBD        0 Dragon 2.0
#> 16                         TBD        0 Dragon 2.0
#> 17                         TBD        0 Dragon 2.0
#> 18                         TBD        0 Dragon 2.0
#>                                                                                                        details
#> 1                                                                         Reentered after three weeks in orbit
#> 2                                                                                      First Dragon soacecraft
#> 3                                                       First of twenty missions flown under the CRS1 contract
#> 4                                                                                                         <NA>
#> 5                                                                                    First Dragon v1.1 capsule
#> 6                                                                            First Dragon capsule to be reused
#> 7                                                                                                         <NA>
#> 8                                                                           Second Dragon capsule to be reused
#> 9                                                                  Destroyed on impact after F9 launch failure
#> 10                                                                   Successful Dragon RTF after CRS-7 failure
#> 11                                                                     In storage at Port of LA as of 11/27/16
#> 12                                                                                                        <NA>
#> 13                                                                        The last newly manufactured Dragon 1
#> 14 Pressure vessel used for Dragon 2 structural testing. Rumored to be repurposed for first Red Dragon Mission
#> 15                          Capsule used to qualify Dragon 2's environmental control and life support systems.
#> 16                                                       Rumored to be used for Inflight Abort Test after DM-1
#> 17                                                                   Currently in construction for use in DM-2
#> 18                              In construction for use in first mission in contract under the CCtCap contract

## get details about all active capsules
get_SpaceX_parts(type = "caps", status = "active")

#>    capsule_serial capsule_id status      original_launch
#> 1            C106    dragon1 active           2014-09-21
#> 2            C108    dragon1 active           2015-04-14
#> 3            C111    dragon1 active           2016-07-18
#> 4            C112    dragon1 active           2017-02-19
#> 5            C113    dragon1 active 2017-08-14T16:31:00Z
#> 6            C201    dragon2 active                  TBD
#> 7            C202    dragon2 active                  TBD
#> 8            C203    dragon2 active                  TBD
#> 9            C204    dragon2 active                  TBD
#> 10           C205    dragon2 active                  TBD
#>                       missions landings       type
#> 1  SpaceX CRS-4, SpaceX CRS-11        2 Dragon 1.1
#> 2  SpaceX CRS-6, SpaceX CRS-13        2 Dragon 1.1
#> 3                 SpaceX CRS-9        1 Dragon 1.1
#> 4                SpaceX CRS-10        1 Dragon 1.1
#> 5                SpaceX CRS-12        1 Dragon 1.1
#> 6                          TBD        0 Dragon 2.0
#> 7                          TBD        0 Dragon 2.0
#> 8                          TBD        0 Dragon 2.0
#> 9                          TBD        0 Dragon 2.0
#> 10                         TBD        0 Dragon 2.0
#>                                                                                                        details
#> 1                                                                            First Dragon capsule to be reused
#> 2                                                                           Second Dragon capsule to be reused
#> 3                                                                      In storage at Port of LA as of 11/27/16
#> 4                                                                                                         <NA>
#> 5                                                                         The last newly manufactured Dragon 1
#> 6  Pressure vessel used for Dragon 2 structural testing. Rumored to be repurposed for first Red Dragon Mission
#> 7                           Capsule used to qualify Dragon 2's environmental control and life support systems.
#> 8                                                        Rumored to be used for Inflight Abort Test after DM-1
#> 9                                                                    Currently in construction for use in DM-2
#> 10                              In construction for use in first mission in contract under the CCtCap contract

## get details about core B1041
get_SpaceX_parts(type = "cores", name = "B1041")

#> $core_serial
#> [1] "B1041"
#> 
#> $status
#> [1] "active"
#> 
#> $original_launch
#> [1] "2017-10-09T12:37:00Z"
#> 
#> $missions
#> [1] "Iridium NEXT 21-30"
#> 
#> $rtls_attempt
#> [1] FALSE
#> 
#> $rtls_landings
#> [1] 0
#> 
#> $asds_attempt
#> [1] TRUE
#> 
#> $asds_landings
#> [1] 1
#> 
#> $water_landing
#> [1] FALSE
#> 
#> $details
#> [1] "In refurbishment"
#>