GET info about SpaceX rockets

GET info about SpaceX rockets

get_SpaceX_rockets(...)

Arguments

...
: Query string to API. See more in details.

Value

data.frame with parsed data from SpaceX API.

Examples

## get details about all rockets
get_SpaceX_rockets()

#>   id active stages boosters cost_per_launch success_rate_pct first_flight
#> 1  1  FALSE      2        0         6700000               40   2006-03-24
#> 2  2   TRUE      2        0        50000000               97   2010-06-04
#> 3  3   TRUE      2        2        90000000              100   2018-02-06
#> 4  4  FALSE      2        0         7000000                0   2021-12-01
#>                            country company height.meters height.feet
#> 1 Republic of the Marshall Islands  SpaceX         22.25        73.0
#> 2                    United States  SpaceX         70.00       229.6
#> 3                    United States  SpaceX         70.00       229.6
#> 4                    United States  SpaceX        118.00       387.0
#>   diameter.meters diameter.feet mass.kg mass.lb
#> 1            1.68           5.5   30146   66460
#> 2            3.70          12.0  549054 1207920
#> 3           12.20          39.9 1420788 3125735
#> 4            9.00          30.0 2943000 1335000
#>                                                                                                                                         payload_weights
#> 1                                                                                                                        leo, Low Earth Orbit, 450, 992
#> 2                                     leo, gto, mars, Low Earth Orbit, Geosynchronous Transfer Orbit, Mars Orbit, 22800, 8300, 4020, 50265, 18300, 8860
#> 3 leo, gto, mars, pluto, Low Earth Orbit, Geosynchronous Transfer Orbit, Mars Orbit, Pluto Orbit, 63800, 26700, 16800, 3500, 140660, 58860, 37040, 7720
#> 4                                              leo, mars, moon, Low Earth Orbit, Mars Orbit, Moon Orbit, 150000, 100000, 100000, 330000, 220000, 220000
#>   first_stage.reusable first_stage.engines first_stage.fuel_amount_tons
#> 1                FALSE                   1                         44.3
#> 2                 TRUE                   9                        385.0
#> 3                 TRUE                  27                       1155.0
#> 4                 TRUE                  37                       3300.0
#>   first_stage.burn_time_sec first_stage.thrust_sea_level.kN
#> 1                       169                             420
#> 2                       162                            7607
#> 3                       162                           22819
#> 4                        NA                          128000
#>   first_stage.thrust_sea_level.lbf first_stage.thrust_vacuum.kN
#> 1                            94000                          480
#> 2                          1710000                         8227
#> 3                          5130000                        24681
#> 4                         28775544                       138000
#>   first_stage.thrust_vacuum.lbf first_stage.cores second_stage.reusable
#> 1                        110000                NA                 FALSE
#> 2                       1849500                NA                 FALSE
#> 3                       5548500                 3                 FALSE
#> 4                      31023634                NA                  TRUE
#>   second_stage.engines second_stage.fuel_amount_tons second_stage.burn_time_sec
#> 1                    1                          3.38                        378
#> 2                    1                         90.00                        397
#> 3                    1                         90.00                        397
#> 4                    6                       1200.00                         NA
#>   second_stage.thrust.kN second_stage.thrust.lbf second_stage.payloads.option_1
#> 1                     31                    7000              composite fairing
#> 2                    934                  210000                         dragon
#> 3                    934                  210000                         dragon
#> 4                   1957                  440000                      Spaceship
#>   second_stage.payloads.composite_fairing.height.meters
#> 1                                                   3.5
#> 2                                                  13.1
#> 3                                                  13.1
#> 4                                                    NA
#>   second_stage.payloads.composite_fairing.height.feet
#> 1                                                11.5
#> 2                                                43.0
#> 3                                                43.0
#> 4                                                  NA
#>   second_stage.payloads.composite_fairing.diameter.meters
#> 1                                                     1.5
#> 2                                                     5.2
#> 3                                                     5.2
#> 4                                                      NA
#>   second_stage.payloads.composite_fairing.diameter.feet
#> 1                                                   4.9
#> 2                                                  17.1
#> 3                                                  17.1
#> 4                                                    NA
#>   second_stage.payloads.option_2 engines.number engines.type engines.version
#> 1                           <NA>              1       merlin              1C
#> 2              composite fairing              9       merlin             1D+
#> 3              composite fairing             27       merlin             1D+
#> 4              composite fairing             37       raptor                
#>   engines.layout engines.isp.sea_level engines.isp.vacuum
#> 1         single                   267                304
#> 2        octaweb                   288                312
#> 3        octaweb                   288                312
#> 4           <NA>                   330                380
#>   engines.engine_loss_max engines.propellant_1 engines.propellant_2
#> 1                       0        liquid oxygen        RP-1 kerosene
#> 2                       2        liquid oxygen        RP-1 kerosene
#> 3                       6        liquid oxygen        RP-1 kerosene
#> 4                      NA        liquid oxygen       liquid methane
#>   engines.thrust_sea_level.kN engines.thrust_sea_level.lbf
#> 1                         420                        94000
#> 2                         845                       190000
#> 3                         845                       190000
#> 4                        1780                       400000
#>   engines.thrust_vacuum.kN engines.thrust_vacuum.lbf engines.thrust_to_weight
#> 1                      480                    110000                     96.0
#> 2                      914                    205500                    180.1
#> 3                      914                    205500                    180.1
#> 4                     1960                    440000                    107.0
#>   landing_legs.number landing_legs.material
#> 1                   0                  <NA>
#> 2                   4          carbon fiber
#> 3                  12          carbon fiber
#> 4                   6       stainless steel
#>                                                                                                                                                                                                                                                                                                                                                                                                                flickr_images
#> 1                                                                                                                          https://www.spacex.com/sites/spacex/files/styles/media_gallery_large/public/2009_-_01_liftoff_south_full_wide_ro8a1280_edit.jpg?itok=8loiSGt1, https://www.spacex.com/sites/spacex/files/styles/media_gallery_large/public/2009_-_02_default_liftoff_west_full_wide_nn6p2062_xl.jpg?itok=p776nHsM
#> 2                     https://farm1.staticflickr.com/929/28787338307_3453a11a77_b.jpg, https://farm4.staticflickr.com/3955/32915197674_eee74d81bb_b.jpg, https://farm1.staticflickr.com/293/32312415025_6841e30bf1_b.jpg, https://farm1.staticflickr.com/623/23660653516_5b6cb301d1_b.jpg, https://farm6.staticflickr.com/5518/31579784413_d853331601_b.jpg, https://farm1.staticflickr.com/745/32394687645_a9c54a34ef_b.jpg
#> 3                                                                                                                                                     https://farm5.staticflickr.com/4599/38583829295_581f34dd84_b.jpg, https://farm5.staticflickr.com/4645/38583830575_3f0f7215e6_b.jpg, https://farm5.staticflickr.com/4696/40126460511_b15bf84c85_b.jpg, https://farm5.staticflickr.com/4711/40126461411_aabc643fd8_b.jpg
#> 4 https://live.staticflickr.com/65535/48954138962_ee541e6755_b.jpg, https://www.spacex.com/sites/spacex/files/starship_cloud_launch1_moved_morechrome.jpg, https://www.spacex.com/sites/spacex/files/starship_uses.jpg, https://live.staticflickr.com/65535/48953946911_e60c5bcc5c_b.jpg, https://live.staticflickr.com/65535/48954138922_9c42173f08_b.jpg, https://live.staticflickr.com/65535/48953947006_313f01ec93_b.jpg
#>                                       wikipedia
#> 1        https://en.wikipedia.org/wiki/Falcon_1
#> 2        https://en.wikipedia.org/wiki/Falcon_9
#> 3    https://en.wikipedia.org/wiki/Falcon_Heavy
#> 4 https://en.wikipedia.org/wiki/SpaceX_Starship
#>                                                                                                                                                                                                                                                                                                       description
#> 1                                                                 The Falcon 1 was an expendable launch system privately developed and manufactured by SpaceX during 2006-2009. On 28 September 2008, Falcon 1 became the first privately-developed liquid-fuel launch vehicle to go into orbit around the Earth.
#> 2                                                                                                                                                      Falcon 9 is a two-stage rocket designed and manufactured by SpaceX for the reliable and safe transport of satellites and the Dragon spacecraft into orbit.
#> 3                 With the ability to lift into orbit over 54 metric tons (119,000 lb)--a mass equivalent to a 737 jetliner loaded with passengers, crew, luggage and fuel--Falcon Heavy can lift more than twice the payload of the next closest operational vehicle, the Delta IV Heavy, at one-third the cost.
#> 4 Starship and Super Heavy Rocket represent a fully reusable transportation system designed to service all Earth orbit needs as well as the Moon and Mars. This two-stage vehicle — composed of the Super Heavy rocket (booster) and Starship (ship) — will eventually replace Falcon 9, Falcon Heavy and Dragon.
#>     rocket_id  rocket_name rocket_type
#> 1     falcon1     Falcon 1      rocket
#> 2     falcon9     Falcon 9      rocket
#> 3 falconheavy Falcon Heavy      rocket
#> 4    starship     Starship      rocket

## get details about "falcon1" rocket
get_SpaceX_rockets(rocket_id = "falcon1")

#>   id active stages boosters cost_per_launch success_rate_pct first_flight
#> 1  1  FALSE      2        0         6700000               40   2006-03-24
#> 2  2   TRUE      2        0        50000000               97   2010-06-04
#> 3  3   TRUE      2        2        90000000              100   2018-02-06
#> 4  4  FALSE      2        0         7000000                0   2021-12-01
#>                            country company height.meters height.feet
#> 1 Republic of the Marshall Islands  SpaceX         22.25        73.0
#> 2                    United States  SpaceX         70.00       229.6
#> 3                    United States  SpaceX         70.00       229.6
#> 4                    United States  SpaceX        118.00       387.0
#>   diameter.meters diameter.feet mass.kg mass.lb
#> 1            1.68           5.5   30146   66460
#> 2            3.70          12.0  549054 1207920
#> 3           12.20          39.9 1420788 3125735
#> 4            9.00          30.0 2943000 1335000
#>                                                                                                                                         payload_weights
#> 1                                                                                                                        leo, Low Earth Orbit, 450, 992
#> 2                                     leo, gto, mars, Low Earth Orbit, Geosynchronous Transfer Orbit, Mars Orbit, 22800, 8300, 4020, 50265, 18300, 8860
#> 3 leo, gto, mars, pluto, Low Earth Orbit, Geosynchronous Transfer Orbit, Mars Orbit, Pluto Orbit, 63800, 26700, 16800, 3500, 140660, 58860, 37040, 7720
#> 4                                              leo, mars, moon, Low Earth Orbit, Mars Orbit, Moon Orbit, 150000, 100000, 100000, 330000, 220000, 220000
#>   first_stage.reusable first_stage.engines first_stage.fuel_amount_tons
#> 1                FALSE                   1                         44.3
#> 2                 TRUE                   9                        385.0
#> 3                 TRUE                  27                       1155.0
#> 4                 TRUE                  37                       3300.0
#>   first_stage.burn_time_sec first_stage.thrust_sea_level.kN
#> 1                       169                             420
#> 2                       162                            7607
#> 3                       162                           22819
#> 4                        NA                          128000
#>   first_stage.thrust_sea_level.lbf first_stage.thrust_vacuum.kN
#> 1                            94000                          480
#> 2                          1710000                         8227
#> 3                          5130000                        24681
#> 4                         28775544                       138000
#>   first_stage.thrust_vacuum.lbf first_stage.cores second_stage.reusable
#> 1                        110000                NA                 FALSE
#> 2                       1849500                NA                 FALSE
#> 3                       5548500                 3                 FALSE
#> 4                      31023634                NA                  TRUE
#>   second_stage.engines second_stage.fuel_amount_tons second_stage.burn_time_sec
#> 1                    1                          3.38                        378
#> 2                    1                         90.00                        397
#> 3                    1                         90.00                        397
#> 4                    6                       1200.00                         NA
#>   second_stage.thrust.kN second_stage.thrust.lbf second_stage.payloads.option_1
#> 1                     31                    7000              composite fairing
#> 2                    934                  210000                         dragon
#> 3                    934                  210000                         dragon
#> 4                   1957                  440000                      Spaceship
#>   second_stage.payloads.composite_fairing.height.meters
#> 1                                                   3.5
#> 2                                                  13.1
#> 3                                                  13.1
#> 4                                                    NA
#>   second_stage.payloads.composite_fairing.height.feet
#> 1                                                11.5
#> 2                                                43.0
#> 3                                                43.0
#> 4                                                  NA
#>   second_stage.payloads.composite_fairing.diameter.meters
#> 1                                                     1.5
#> 2                                                     5.2
#> 3                                                     5.2
#> 4                                                      NA
#>   second_stage.payloads.composite_fairing.diameter.feet
#> 1                                                   4.9
#> 2                                                  17.1
#> 3                                                  17.1
#> 4                                                    NA
#>   second_stage.payloads.option_2 engines.number engines.type engines.version
#> 1                           <NA>              1       merlin              1C
#> 2              composite fairing              9       merlin             1D+
#> 3              composite fairing             27       merlin             1D+
#> 4              composite fairing             37       raptor                
#>   engines.layout engines.isp.sea_level engines.isp.vacuum
#> 1         single                   267                304
#> 2        octaweb                   288                312
#> 3        octaweb                   288                312
#> 4           <NA>                   330                380
#>   engines.engine_loss_max engines.propellant_1 engines.propellant_2
#> 1                       0        liquid oxygen        RP-1 kerosene
#> 2                       2        liquid oxygen        RP-1 kerosene
#> 3                       6        liquid oxygen        RP-1 kerosene
#> 4                      NA        liquid oxygen       liquid methane
#>   engines.thrust_sea_level.kN engines.thrust_sea_level.lbf
#> 1                         420                        94000
#> 2                         845                       190000
#> 3                         845                       190000
#> 4                        1780                       400000
#>   engines.thrust_vacuum.kN engines.thrust_vacuum.lbf engines.thrust_to_weight
#> 1                      480                    110000                     96.0
#> 2                      914                    205500                    180.1
#> 3                      914                    205500                    180.1
#> 4                     1960                    440000                    107.0
#>   landing_legs.number landing_legs.material
#> 1                   0                  <NA>
#> 2                   4          carbon fiber
#> 3                  12          carbon fiber
#> 4                   6       stainless steel
#>                                                                                                                                                                                                                                                                                                                                                                                                                flickr_images
#> 1                                                                                                                          https://www.spacex.com/sites/spacex/files/styles/media_gallery_large/public/2009_-_01_liftoff_south_full_wide_ro8a1280_edit.jpg?itok=8loiSGt1, https://www.spacex.com/sites/spacex/files/styles/media_gallery_large/public/2009_-_02_default_liftoff_west_full_wide_nn6p2062_xl.jpg?itok=p776nHsM
#> 2                     https://farm1.staticflickr.com/929/28787338307_3453a11a77_b.jpg, https://farm4.staticflickr.com/3955/32915197674_eee74d81bb_b.jpg, https://farm1.staticflickr.com/293/32312415025_6841e30bf1_b.jpg, https://farm1.staticflickr.com/623/23660653516_5b6cb301d1_b.jpg, https://farm6.staticflickr.com/5518/31579784413_d853331601_b.jpg, https://farm1.staticflickr.com/745/32394687645_a9c54a34ef_b.jpg
#> 3                                                                                                                                                     https://farm5.staticflickr.com/4599/38583829295_581f34dd84_b.jpg, https://farm5.staticflickr.com/4645/38583830575_3f0f7215e6_b.jpg, https://farm5.staticflickr.com/4696/40126460511_b15bf84c85_b.jpg, https://farm5.staticflickr.com/4711/40126461411_aabc643fd8_b.jpg
#> 4 https://live.staticflickr.com/65535/48954138962_ee541e6755_b.jpg, https://www.spacex.com/sites/spacex/files/starship_cloud_launch1_moved_morechrome.jpg, https://www.spacex.com/sites/spacex/files/starship_uses.jpg, https://live.staticflickr.com/65535/48953946911_e60c5bcc5c_b.jpg, https://live.staticflickr.com/65535/48954138922_9c42173f08_b.jpg, https://live.staticflickr.com/65535/48953947006_313f01ec93_b.jpg
#>                                       wikipedia
#> 1        https://en.wikipedia.org/wiki/Falcon_1
#> 2        https://en.wikipedia.org/wiki/Falcon_9
#> 3    https://en.wikipedia.org/wiki/Falcon_Heavy
#> 4 https://en.wikipedia.org/wiki/SpaceX_Starship
#>                                                                                                                                                                                                                                                                                                       description
#> 1                                                                 The Falcon 1 was an expendable launch system privately developed and manufactured by SpaceX during 2006-2009. On 28 September 2008, Falcon 1 became the first privately-developed liquid-fuel launch vehicle to go into orbit around the Earth.
#> 2                                                                                                                                                      Falcon 9 is a two-stage rocket designed and manufactured by SpaceX for the reliable and safe transport of satellites and the Dragon spacecraft into orbit.
#> 3                 With the ability to lift into orbit over 54 metric tons (119,000 lb)--a mass equivalent to a 737 jetliner loaded with passengers, crew, luggage and fuel--Falcon Heavy can lift more than twice the payload of the next closest operational vehicle, the Delta IV Heavy, at one-third the cost.
#> 4 Starship and Super Heavy Rocket represent a fully reusable transportation system designed to service all Earth orbit needs as well as the Moon and Mars. This two-stage vehicle — composed of the Super Heavy rocket (booster) and Starship (ship) — will eventually replace Falcon 9, Falcon Heavy and Dragon.
#>     rocket_id  rocket_name rocket_type
#> 1     falcon1     Falcon 1      rocket
#> 2     falcon9     Falcon 9      rocket
#> 3 falconheavy Falcon Heavy      rocket
#> 4    starship     Starship      rocket